refactoring and cleaning main workflow
