--- src/Svg.tsx
@@ -1,557 +1,0 @@
-import * as React from 'react'
-import { SVGProps } from 'react'
-import { usePaletteContext } from './interface/palette/usePaletteContext'
-import { PaletteColorType } from './services/palette'
-
-// const { palette } = usePaletteContext()
-const lightBlockHex = '#E6DAFD'
-const darkBlobGradient = '#C4C4C4'
-const blueShirtandHair = '#829EFF'
-const purpleTabletAndSwatches = '#5624C2'
-const blackBorded = '#210E10'
-// svg fill colors props
-interface SvgCmpProps extends SVGProps<SVGSVGElement> {
-   lightBlockHex: string
-   primary: PaletteColorType
-   secondary: PaletteColorType
-}
-
-const SvgComponent = (props: SvgCmpProps) => (
-   <svg width={537} height={526} fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
-      <g filter="url(#a)">
-         {/* blob in the back */}
-         <path
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M245.962 107.408c27.071 11.929 55.129 15.37 82.047 27.642 40.065 18.266 108.904 17.597 113.291 61.41 4.623 46.166-69.489 57.462-97.699 94.297-18.143 23.689-12.629 60.687-35.591 79.742-24.567 20.388-58.752 18.748-90.138 24.595-46.105 8.589-99.075 46.31-136.125 17.555-36.346-28.208-18.99-88.217-13.83-133.935 3.804-33.693 32.721-57.868 39.24-91.142 9.642-49.222-36.1-121.523 7.335-146.608 43.587-25.172 85.41 46.147 131.47 66.444Z"
-            fill={props.lightBlockHex}
-         />
-
-         {/* big blob in front */}
-         <path
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M298.949 15.17c34.499 4.775 50.985 46.045 82.08 61.732 38.879 19.613 94.499 7.028 121.672 41.055 26.43 33.096 15.512 82.399 10.5 124.456-4.936 41.426-17.234 81.152-39.227 116.604-22.969 37.027-49.537 75.536-89.998 91.708-40.044 16.005-86.073 8.425-127.07-4.949-36.642-11.952-58.314-48.063-92.129-66.555-39.919-21.83-96.983-14.853-123.864-51.56-26.621-36.352-32.253-92.763-10.871-132.423 22.124-41.038 83.261-38.807 119.198-68.508 24.011-19.844 28.814-55.848 53.792-74.462 27.964-20.838 61.372-41.879 95.917-37.098Z"
-            fill="#C4C4C4"
-         />
-         <path
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M298.949 15.17c34.499 4.775 50.985 46.045 82.08 61.732 38.879 19.613 94.499 7.028 121.672 41.055 26.43 33.096 15.512 82.399 10.5 124.456-4.936 41.426-17.234 81.152-39.227 116.604-22.969 37.027-49.537 75.536-89.998 91.708-40.044 16.005-86.073 8.425-127.07-4.949-36.642-11.952-58.314-48.063-92.129-66.555-39.919-21.83-96.983-14.853-123.864-51.56-26.621-36.352-32.253-92.763-10.871-132.423 22.124-41.038 83.261-38.807 119.198-68.508 24.011-19.844 28.814-55.848 53.792-74.462 27.964-20.838 61.372-41.879 95.917-37.098Z"
-            fill="url(#b)"
-         />
-         <path
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M298.949 15.17c34.499 4.775 50.985 46.045 82.08 61.732 38.879 19.613 94.499 7.028 121.672 41.055 26.43 33.096 15.512 82.399 10.5 124.456-4.936 41.426-17.234 81.152-39.227 116.604-22.969 37.027-49.537 75.536-89.998 91.708-40.044 16.005-86.073 8.425-127.07-4.949-36.642-11.952-58.314-48.063-92.129-66.555-39.919-21.83-96.983-14.853-123.864-51.56-26.621-36.352-32.253-92.763-10.871-132.423 22.124-41.038 83.261-38.807 119.198-68.508 24.011-19.844 28.814-55.848 53.792-74.462 27.964-20.838 61.372-41.879 95.917-37.098Z"
-            fill={props.primary.shade[900].hex}
-         />
-         <path
-            d="m277.566 277.786 1.089-25.95a4.955 4.955 0 0 1 .23-1.904 4.893 4.893 0 0 1 4.308-3.357l1.723-.137a4.95 4.95 0 0 1 3.592 1.171 4.95 4.95 0 0 1 1.716 3.365l2.449 28.808"
-            fill="#fff"
-         />
-         {/* end of blob */}
-         <path
-            d="M292.673 280.915a1.136 1.136 0 0 1-1.135-1.044l-2.449-28.806a3.807 3.807 0 0 0-4.083-3.494l-1.723.137a3.623 3.623 0 0 0-2.586 1.315 3.578 3.578 0 0 0-.908 2.721v.137l-1.088 25.949a1.077 1.077 0 0 1-.351.804c-.11.102-.243.181-.385.229-.142.05-.294.068-.445.056a1.186 1.186 0 0 1-1.088-1.181l1.088-25.857a6.133 6.133 0 0 1 1.453-4.356 5.811 5.811 0 0 1 4.127-2.086l1.726-.136a5.804 5.804 0 0 1 4.399 1.451 5.989 5.989 0 0 1 1.48 1.852 6 6 0 0 1 .653 2.277l2.449 28.807a1.135 1.135 0 0 1-1.042 1.225h-.092Z"
-            fill="#000"
-         />
-         <path
-            d="m443.152 353.865-27.174-44.64s-54.121 36.246-63.603 28.443c-5.626-4.626-49.448-48.859-57.75-58.432a38.405 38.405 0 0 1-3.041-4.083 152.236 152.236 0 0 0-8.165-12.475c-2.949-4.128-19.372-25.178-19.372-25.178a4.355 4.355 0 0 0-6.125-.362l-1.089.952a3.04 3.04 0 0 0-.406 4.173l11.749 18.237-10.344 10.434-13.157-17.918c-1.587-1.816-4.399-4.083-7.031-2.223-2.631 1.86-1.587 5.716-.589 6.985l28.579 37.292s53.624 73.039 70.59 86.332c21.913 16.966 106.928-27.537 106.928-27.537Z"
-            fill="#fff"
-         />
-         <path
-            d="M351.015 386.394c-6.442 0-11.796-1.227-15.516-4.085-16.876-13.291-68.593-83.61-70.816-86.602l-28.535-37.247a7.393 7.393 0 0 1-1.225-5.127 5.137 5.137 0 0 1 .689-1.938 5.133 5.133 0 0 1 1.397-1.509c2.496-1.77 5.627-.908 8.529 2.359a.146.146 0 0 1 .046.09l12.385 16.877 8.756-8.892-11.206-17.375a4.226 4.226 0 0 1 .543-5.716l1.091-.952a5.535 5.535 0 0 1 3.98-1.343 5.535 5.535 0 0 1 3.776 1.842h.047c.679.862 16.467 21.095 19.371 25.178a150.817 150.817 0 0 1 8.256 12.611 35.727 35.727 0 0 0 2.948 3.902c8.801 10.162 52.217 53.894 57.57 58.296 1.996 1.631 9.028 2.313 34.071-11.433 14.516-7.985 28.035-17.012 28.171-17.102.257-.167.563-.231.862-.183.152.036.297.102.421.197.124.094.228.212.306.348l27.175 44.64a.993.993 0 0 1 .134.862 1.09 1.09 0 0 1-.196.419 1.087 1.087 0 0 1-.347.307c-2.948 1.587-60.654 31.576-92.683 31.576ZM239.777 251.247a2.267 2.267 0 0 0-1.453.499 2.627 2.627 0 0 0-1.132 1.906 4.99 4.99 0 0 0 .77 3.448l28.535 37.246c.545.769 53.76 73.083 70.409 86.195 19.779 15.287 94.724-22.093 104.659-27.13l-25.949-42.598c-3.358 2.223-15.017 9.753-27.355 16.558-20.234 11.116-32.212 14.79-36.611 11.16-6.033-4.945-49.813-49.266-57.886-58.566a51.804 51.804 0 0 1-3.177-4.219c-3.131-5.354-5.989-9.437-8.076-12.431-2.858-3.947-18.644-24.18-19.325-25.088a3.266 3.266 0 0 0-4.536-.273l-1.09.999c-.369.324-.6.779-.644 1.27-.042.489.109.978.417 1.361a.146.146 0 0 1 .046.09l11.75 18.193c.148.211.216.47.19.728-.024.257-.14.497-.327.678l-10.342 10.48a1.133 1.133 0 0 1-.417.265 1.109 1.109 0 0 1-.491.052.986.986 0 0 1-.465-.125 1.027 1.027 0 0 1-.353-.328l-13.11-17.875c-.952-1.044-2.495-2.495-4.037-2.495Z"
-            fill="#000"
-         />
-         <path
-            d="m264.137 270.119-15.786-26.266a3.085 3.085 0 0 1 .363-4.175l1.134-.952a4.352 4.352 0 0 1 6.123.362l17.602 25.042"
-            fill="#fff"
-         />
-         <path
-            d="M264.137 271.254a1.219 1.219 0 0 1-.539-.16 1.247 1.247 0 0 1-.413-.385l-15.788-26.221a4.17 4.17 0 0 1-.801-2.928 4.162 4.162 0 0 1 1.393-2.698l1.088-.998a5.451 5.451 0 0 1 1.86-1.078 5.456 5.456 0 0 1 2.133-.283c1.449.106 2.8.774 3.764 1.86 0 .046.046.046.046.09l17.648 25.042a1.146 1.146 0 0 1-.273 1.59 1.139 1.139 0 0 1-1.587-.273l-17.602-24.998a3.039 3.039 0 0 0-.982-.745 3.018 3.018 0 0 0-1.197-.297 3.077 3.077 0 0 0-2.313.816l-1.09.954a1.905 1.905 0 0 0-.227 2.629c.004.037.02.069.046.093l15.832 26.266a1.127 1.127 0 0 1-.406 1.541.909.909 0 0 1-.592.183Z"
-            fill="#000"
-         />
-         <path
-            d="m260.78 275.653-9.344-13.2c-.409-.455-5.672-7.803-7.985-11.75-2.133-3.177-2.858-4.491-.18-6.67 2.675-2.177 5.261.681 6.848 2.495l17.738 23.817"
-            fill="#fff"
-         />
-         <path
-            d="M260.827 276.788a1.068 1.068 0 0 1-.514-.108 1.048 1.048 0 0 1-.394-.345c-.045 0-.045 0-1.588-2.178l-7.757-10.978c-.681-.818-5.806-8.029-8.075-11.885-1.996-2.905-3.539-5.264.044-8.122 3.586-2.858 6.987.952 8.438 2.586.046 0 .046.046.046.092l17.739 23.817a1.135 1.135 0 0 1 .216.84 1.144 1.144 0 0 1-.842.938 1.124 1.124 0 0 1-1.188-.417l-17.694-23.817c-2.496-2.814-3.901-3.403-5.262-2.315-1.361 1.088-1.723 1.996.407 5.173h.046c2.269 3.946 7.486 11.16 7.849 11.613l7.893 11.114 1.543 2.179a1.084 1.084 0 0 1 .211.817 1.089 1.089 0 0 1-.437.724.93.93 0 0 1-.315.204.916.916 0 0 1-.366.068Z"
-            fill="#000"
-         />
-         {/* men shirt and border */}
-         <path
-            d="M319.075 464.015s3.72-74.673 7.214-126.709c3.493-52.034 12.703-68.82 21.731-82.882 22.276-31.122 66.733-25.044 83.11-10.707 17.286 12.202 21.685 44.367 27.538 75.125 15.379 105.068 14.699 146.987 14.699 146.987l-154.292-1.814Z"
-            fill={props.secondary.shade[500].hex}
-         />
-         <path
-            d="m473.368 466.962-154.292-1.814a1.375 1.375 0 0 1-.816-.363 1.168 1.168 0 0 1-.316-.864c.044-.725 3.764-75.262 7.257-126.706 1.588-23.591 4.446-41.874 8.801-55.937 3.947-12.747 8.621-20.55 13.066-27.491h.044c11.525-16.151 27.766-21.141 39.333-22.456 16.74-1.952 35.386 2.812 45.367 11.523 16.059 11.386 21.096 38.605 26.403 67.413.499 2.768.998 5.626 1.542 8.394 12.204 83.201 14.972 132.649 14.745 147.166a1.136 1.136 0 0 1-1.134 1.135Zm-153.065-4.083 151.93 1.77c.091-15.788-2.858-64.693-14.699-145.625-.543-2.815-1.043-5.627-1.588-8.395-4.99-27.128-10.162-55.164-25.494-66.007 0-.045-.047-.045-.093-.091-9.526-8.348-27.491-12.883-43.642-10.977-11.113 1.268-26.675 6.033-37.745 21.502-7.53 11.751-17.828 27.856-21.548 82.339-3.221 47.863-6.668 115.731-7.121 125.484Z"
-            fill={props.primary.shade[900].hex}
-         />
-
-         <path
-            d="M227.21 464.65s-10.072-142.131-7.576-169.897c0-28.399-10.252-38.605-18.6-45.456-32.391-24.18-71.951-17.964-83.973-9.935-26.766 21.504-28.762 60.835-28.762 60.835L72.92 466.464l154.29-1.814Z"
-            fill="#000"
-         />
-         <path
-            d="M72.92 467.599a1.052 1.052 0 0 1-.817-.363 1.039 1.039 0 0 1-.318-.861l15.38-166.268c.045-.679.635-10.252 4.4-22.593 3.539-11.566 10.752-27.717 24.815-39.06a.045.045 0 0 0 .032-.012.047.047 0 0 0 .014-.032c8.256-5.536 23.545-7.668 35.022-7.305 12.884.408 32.074 3.72 50.267 17.285h.044c6.668 5.49 19.009 15.605 19.009 46.364v.092c-2.495 27.355 7.44 168.308 7.576 169.713a1.174 1.174 0 0 1-.319.862c-.104.114-.23.204-.37.266a1.176 1.176 0 0 1-.445.098l-154.29 1.814ZM117.74 240.27c-26.04 20.959-28.263 59.611-28.308 59.973v.09L74.144 465.33l151.841-1.814c-.952-13.837-9.845-142.54-7.486-168.853 0-27.31-9.344-37.29-18.145-44.459-32.846-24.451-71.815-17.103-82.613-9.934Z"
-            fill="#000"
-         />
-         <path
-            d="M174.994 232.92c5.943 1.588 15.153 10.072 12.884 14.88-2.269 4.808-8.665 8.711-29.352 6.622-9.572-1.178-23.227-5.942-28.263-10.252-16.286-13.882 11.115-16.423 11.115-16.423"
-            fill="#000"
-         />
-         <path
-            d="M168.916 256.147a101.37 101.37 0 0 1-10.525-.591c-11.433-1.405-24.362-6.668-28.853-10.525-4.491-3.856-6.306-7.304-5.353-10.388.952-3.085 4.128-5.035 9.526-6.578a46.15 46.15 0 0 1 7.531-1.451 1.18 1.18 0 0 1 1.271 1.042 1.15 1.15 0 0 1-.267.83 1.105 1.105 0 0 1-.349.274c-.133.068-.279.11-.428.12-3.72.363-13.746 2.179-15.107 6.442-.816 2.632 2.042 5.809 4.628 7.985 4.717 3.993 17.874 8.801 27.673 10.026 21.729 2.133 26.493-2.403 28.173-5.987.453-.908.317-2.088-.319-3.493-1.814-3.903-7.576-8.711-11.795-9.845a1.123 1.123 0 0 1-.406-.195 1.1 1.1 0 0 1-.299-.337 1.136 1.136 0 0 1-.111-.876 1.139 1.139 0 0 1 .965-.809c.148-.016.298-.002.441.04 4.946 1.315 11.115 6.486 13.248 11.068.952 1.996 1.042 3.902.363 5.399-2.496 5.308-8.801 7.849-20.007 7.849Z"
-            fill="#fff"
-         />
-         <path
-            d="M109.394 189.142c-11.885 8.12-12.067 17.692-4.264 29.125 7.803 11.432 16.786 14.699 28.671 6.624 11.886-8.076 25.678-31.848 20.143-39.969-5.535-8.12-32.709-3.856-44.55 4.22Z"
-            fill="#829EFF"
-         />
-         <path
-            d="M121.553 230.468c-6.352 0-12.113-3.856-17.375-11.569-4.22-6.215-5.989-11.614-5.444-16.559.544-4.945 3.901-9.98 10.026-14.154 11.613-7.939 39.876-13.065 46.137-3.901 2.767 4.083 1.497 11.795-3.493 21.14-4.492 8.438-11.16 16.423-16.967 20.415-4.582 3.085-8.847 4.628-12.884 4.628Zm18.963-48.769c-1.951 0-4.083.136-6.352.363-9.118 1.089-18.6 4.219-24.134 8.029-5.535 3.811-8.484 7.849-9.028 12.476-.544 4.627 1.134 9.346 5.036 15.062 8.029 11.705 16.377 13.655 27.083 6.306 5.58-3.766 11.931-11.478 16.287-19.553 4.355-8.075 5.852-15.515 3.584-18.827-1.724-2.495-6.216-3.856-12.476-3.856Z"
-            fill="#829EFF"
-         />
-         <path
-            d="m143.555 228.745 11.205-43.371 25.724 4.718-8.212 44.959c1.905.771 6.305 5.081 5.806 7.847-.499 2.767-4.581 4.764-20.868 1.769-16.286-2.994-21.231-7.303-20.641-10.705.408-2.132 5.806-1.906 6.986-5.217Z"
-            fill="#fff"
-         />
-         <path
-            d="M170.911 247.481c-3.357 0-7.894-.591-13.928-1.679-7.667-1.407-13.383-3.175-17.012-5.354-3.629-2.177-4.809-4.991-4.491-6.668.317-1.68 1.905-2.223 3.447-2.858 1.543-.636 3.04-1.225 3.539-2.542l11.206-43.278c.074-.281.25-.524.493-.682.242-.16.535-.224.822-.18l25.723 4.718a1.18 1.18 0 0 1 .908 1.315l-8.076 44.142c2.404 1.541 6.216 5.624 5.625 8.665-.59 3.038-2.903 4.401-8.256 4.401Zm-15.289-60.791-10.978 42.325c0 .047 0 .093-.046.093-.816 2.313-3.039 3.221-4.854 3.9-.771.319-1.996.818-2.041 1.181-.227 1.36.952 2.902 3.402 4.309 3.403 1.996 8.892 3.72 16.287 5.081 18.781 3.447 19.326.499 19.552-.908.227-1.405-3.402-5.853-5.126-6.578a1.179 1.179 0 0 1-.557-.519c-.123-.229-.166-.494-.123-.75l8.029-43.825-23.545-4.309Z"
-            fill="#000"
-         />
-         <path
-            d="M157.21 207.741c4.673 5.535 10.797 9.209 17.784 10.117 18.1 2.313 35.249-14.835 38.244-38.335 2.994-23.499-9.21-44.413-27.311-46.727s-35.204 14.835-38.244 38.334c-1.361 10.752.499 21.368 4.627 29.534"
-            fill="#fff"
-         />
-         <path
-            d="M178.261 219.219a28.445 28.445 0 0 1-3.403-.227c-7.032-.907-13.474-4.536-18.51-10.479a1.145 1.145 0 0 1 .942-1.888c.302.022.583.162.782.391 4.673 5.534 10.57 8.891 17.058 9.708 17.465 2.268 34.069-14.472 36.973-37.336 2.903-22.865-8.847-43.234-26.312-45.457-17.466-2.223-34.025 14.471-36.974 37.336-1.27 10.117.363 20.642 4.537 28.853a1.145 1.145 0 0 1-1.374 1.612 1.146 1.146 0 0 1-.668-.569c-4.355-8.619-6.079-19.643-4.718-30.213 3.085-24.045 20.823-41.692 39.469-39.288 18.645 2.405 31.394 23.954 28.309 48.043-2.904 22.593-18.692 39.514-36.111 39.514Z"
-            fill="#000"
-         />
-         <path
-            d="M189.012 192.77a1.095 1.095 0 0 1-1.044-.726 1.07 1.07 0 0 1 .006-.863 1.099 1.099 0 0 1 .629-.589c1.952-.726 7.123-3.175 7.985-6.079.181-1.043-.182-1.361-1.587-2.268-1.407-.908-4.446-2.813-5.354-7.894-.635-3.674.908-6.578 1.724-8.166.172-.252.309-.527.409-.816a1.131 1.131 0 0 1 1.224-.953 1.177 1.177 0 0 1 1.043 1.27 5.749 5.749 0 0 1-.634 1.588c-.771 1.406-2.042 3.811-1.543 6.714.726 4.083 2.632 5.308 4.355 6.352 1.724 1.043 3.085 1.996 2.586 4.672v.091c-1.315 4.537-8.574 7.304-9.39 7.621l-.409.046ZM174.994 217.857s-3.493-.545-5.171-1.225a44.757 44.757 0 0 1-6.08-2.994c-.907-.635-6.532-5.898-6.532-5.898a20.136 20.136 0 0 0 7.395 9.345c5.489 3.584 9.708 4.492 9.708 4.492l.68-3.72Z"
-            fill="#000"
-         />
-         <path
-            d="M174.313 222.711h-.227c-.181-.045-4.491-1.043-10.071-4.673a21.001 21.001 0 0 1-7.894-9.935 1.182 1.182 0 0 1 .454-1.315c.218-.133.472-.191.726-.166.253.024.492.13.68.302 2.133 1.996 5.807 5.353 6.442 5.761.636.409 4.083 2.223 5.807 2.904 1.724.68 4.9 1.134 4.945 1.179a1.045 1.045 0 0 1 .771.454 1.136 1.136 0 0 1 .182.862l-.681 3.72a1.121 1.121 0 0 1-.499.726 1.045 1.045 0 0 1-.635.181Zm-10.842-7.939c.55.503 1.141.959 1.769 1.361a33.942 33.942 0 0 0 8.211 4.038l.227-1.407a27.891 27.891 0 0 1-4.264-1.043 48.374 48.374 0 0 1-5.943-2.949ZM153.898 183.878a8.758 8.758 0 0 1-5.626 5.535l-.589-1.951c2.223.181 4.4-.499 6.215-3.584Z"
-            fill="#000"
-         />
-         <path
-            d="M148.273 190.547a.907.907 0 0 1-.499-.136 1.132 1.132 0 0 1-.59-.635l-.589-1.951a1.085 1.085 0 0 1 .181-1.043 1.174 1.174 0 0 1 .998-.454c1.588.136 3.448-.136 5.126-3.039a1.182 1.182 0 0 1 1.452-.454 1.09 1.09 0 0 1 .59 1.406 9.885 9.885 0 0 1-6.306 6.261l-.363.045Z"
-            fill="#000"
-         />
-         <path
-            d="M154.67 172.308c-1.77-5.081-5.989-8.347-10.026-7.531-4.038.817-7.214 6.76-5.943 12.975 1.27 6.215 5.988 10.525 10.57 9.618a7.13 7.13 0 0 0 4.627-3.494"
-            fill="#fff"
-         />
-         <path
-            d="M148.001 188.597c-2.087 0-4.174-.862-6.034-2.495a15.154 15.154 0 0 1-4.4-8.121c-1.406-6.85 1.678-13.247 6.85-14.29 5.172-1.044 9.3 2.541 11.296 8.257a1.131 1.131 0 0 1-.68 1.451 1.18 1.18 0 0 1-1.452-.68c-1.633-4.673-5.263-7.531-8.71-6.805-3.448.726-6.216 5.988-5.081 11.614a13.105 13.105 0 0 0 3.72 6.895c1.724 1.543 3.674 2.178 5.534 1.815a5.853 5.853 0 0 0 3.856-2.949 1.18 1.18 0 0 1 1.917-.069 1.08 1.08 0 0 1 .034 1.249 8.208 8.208 0 0 1-5.353 3.992c-.494.09-.995.136-1.497.136Z"
-            fill="#000"
-         />
-         <path
-            d="M204.664 133.023c-10.208-6.396-21.505-7.167-27.085-2.404-5.58-4.038-16.014-1.633-24.497 5.988-8.484 7.622-13.519 22.547-8.438 28.173 1.406 1.542 3.312 2.268 5.58 2.359 1.905 2.359 3.357 5.489 4.446 8.937 1.905-3.584 5.761-9.708 9.572-13.61a49.238 49.238 0 0 0 7.304-5.444 33.918 33.918 0 0 0 6.532-7.984 35.928 35.928 0 0 0 10.525 9.753c5.307 3.312 13.475 4.492 19.781 5.127 5.034.544 5.307 8.801 5.307 8.801s5.127-7.168 5.853-16.151c.725-8.982-3.267-16.331-14.88-23.545Z"
-            fill="#829EFF"
-         />
-         <path
-            d="m154.397 178.977-.817-2.586a26.383 26.383 0 0 0-3.946-8.166c-2.45-.181-4.401-1.088-5.807-2.676-5.58-6.17-1.316-20.914 8.529-29.806 8.347-7.53 18.963-10.252 25.177-6.532 6.215-4.537 17.694-3.448 27.766 2.858 11.16 6.941 16.194 14.971 15.379 24.588-.818 9.618-5.853 16.423-6.035 16.74l-1.951 2.722-.136-3.357c0-.045-.317-7.304-4.309-7.712-5.989-.59-14.654-1.815-20.234-5.308a37.397 37.397 0 0 1-9.845-8.71 33.214 33.214 0 0 1-5.897 6.805 44.553 44.553 0 0 1-7.304 5.489c-3.81 3.992-7.667 10.207-9.3 13.292l-1.27 2.359Zm16.468-49.177c-5.172 0-11.568 2.722-17.012 7.622-8.665 7.848-12.975 21.503-8.348 26.585 1.089 1.224 2.722 1.905 4.764 1.996h.544l.318.362a25.996 25.996 0 0 1 3.765 6.896c2.223-3.765 5.399-8.438 8.574-11.659l.182-.181a46.868 46.868 0 0 0 7.122-5.263 31.824 31.824 0 0 0 6.352-7.712l.906-1.497.998 1.452a34.714 34.714 0 0 0 10.208 9.436c5.171 3.221 13.475 4.4 19.235 4.99 3.674.363 5.218 3.856 5.899 6.624 1.495-2.813 3.584-7.577 4.037-12.975.725-8.665-3.947-16.014-14.335-22.502-9.619-5.988-20.462-7.032-25.723-2.495l-.682.59-.725-.545c-1.633-1.179-3.765-1.724-6.079-1.724Z"
-            fill="#829EFF"
-         />
-         <path
-            d="M148.953 184.057a.915.915 0 0 1-.499-.136c-3.266-1.543-6.759-4.991-4.854-12.476a1.15 1.15 0 1 1 2.223.59c-1.27 4.854-.091 8.075 3.584 9.844a1.089 1.089 0 0 1 .545 1.497 1.14 1.14 0 0 1-.999.681Z"
-            fill="#000"
-         />
-         <path
-            d="M186.563 203.112c-.998 1.906-3.72 2.405-6.126 1.18-2.405-1.225-3.493-3.811-2.495-5.671.998-1.86 2.948-.453 5.308.771 2.359 1.225 4.265 1.86 3.313 3.72Z"
-            fill="#fff"
-         />
-         <path
-            d="M183.115 206.108a6.941 6.941 0 0 1-3.222-.817 6.673 6.673 0 0 1-2.994-3.13 4.627 4.627 0 0 1 0-4.083 2.906 2.906 0 0 1 1.633-1.406c1.316-.408 2.721.363 4.355 1.27l.908.454.862.454c1.497.771 2.768 1.406 3.175 2.631.108.362.14.742.094 1.117a2.738 2.738 0 0 1-.365 1.06 4.682 4.682 0 0 1-3.357 2.314 4.5 4.5 0 0 1-1.089.136Zm-3.812-7.258h-.136a.683.683 0 0 0-.227.272 2.413 2.413 0 0 0 0 2.087 4.541 4.541 0 0 0 2.043 2.086c.862.456 1.85.615 2.812.454a2.26 2.26 0 0 0 .998-.36c.301-.197.549-.462.726-.774.136-.272.182-.408.136-.454-.046-.045-1.179-.862-1.996-1.27l-.952-.454-.908-.499c-.862-.499-1.997-1.088-2.496-1.088ZM203.767 175.492c.854-2.383.493-4.691-.803-5.155-1.299-.465-3.041 1.09-3.895 3.473-.854 2.382-.493 4.69.804 5.155 1.299.464 3.04-1.091 3.894-3.473ZM181.688 168.182c.854-2.382.493-4.69-.804-5.155-1.297-.464-3.041 1.09-3.894 3.473-.853 2.383-.493 4.69.805 5.155 1.297.465 3.041-1.09 3.893-3.473Z"
-            fill="#000"
-         />
-         <path
-            d="M178.125 154.797a1.133 1.133 0 0 1-1.135-1.134v-13.655a1.134 1.134 0 0 1 2.269 0v13.655a1.135 1.135 0 0 1-1.134 1.134ZM100.457 210.008a1.174 1.174 0 0 1-1.044-.68 1.139 1.139 0 0 1 .224-1.245c.104-.107.228-.193.366-.252l16.967-7.44a1.081 1.081 0 0 1 1.254.212c.106.107.189.236.243.377a1.054 1.054 0 0 1 .033.86 1.04 1.04 0 0 1-.623.592l-16.967 7.486-.453.09ZM106.536 221.351a1.219 1.219 0 0 1-.998-.544 1.183 1.183 0 0 1 .408-1.588l8.665-5.172a1.135 1.135 0 0 1 1.543.408 1.083 1.083 0 0 1 .151.85 1.085 1.085 0 0 1-.514.693l-8.711 5.172a.856.856 0 0 1-.544.181ZM115.927 228.156a1.032 1.032 0 0 1-.771-.317 1.135 1.135 0 0 1-.045-1.587l20.188-21.096a1.086 1.086 0 0 1 1.588-.045 1.134 1.134 0 0 1 .045 1.587l-20.188 21.141a1.045 1.045 0 0 1-.817.317ZM246.898 303.282l-8.346-24.589a4.942 4.942 0 0 0-2.153-3.087 4.938 4.938 0 0 0-3.699-.679l-1.68.363a4.908 4.908 0 0 0-3.131 2.128 4.9 4.9 0 0 0-.679 3.724l5.806 28.309"
-            fill="#fff"
-         />
-         <path
-            d="M233.017 310.587a1.188 1.188 0 0 1-.724-.259 1.19 1.19 0 0 1-.411-.649l-5.806-28.309a6.123 6.123 0 0 1 4.674-7.213l1.724-.363a6.083 6.083 0 0 1 7.167 4.628l8.348 24.498a1.13 1.13 0 0 1-.728 1.405 1.036 1.036 0 0 1-1.2-.292 1.018 1.018 0 0 1-.205-.387l-8.394-24.589a.126.126 0 0 0-.044-.09 3.757 3.757 0 0 0-1.627-2.417 3.768 3.768 0 0 0-2.865-.533l-1.679.362a3.866 3.866 0 0 0-2.385 1.648 3.867 3.867 0 0 0-.563 2.844l5.806 28.355c.034.146.038.297.014.445a1.132 1.132 0 0 1-.876.916h-.226Z"
-            fill="#000"
-         />
-         {/* tablet fill */}
-         <path
-            d="M312.671 194.882 262.7 176.989c-7.761-2.779-16.305 1.26-19.085 9.021l-27.273 76.174c-2.78 7.76 1.259 16.305 9.019 19.083l49.972 17.892c7.761 2.78 16.305-1.258 19.083-9.019l27.275-76.175c2.78-7.761-1.259-16.305-9.02-19.083Z"
-            fill="#5624C2"
-         />
-         {/* tablet border */}
-         <path
-            d="M280.061 301.376a16.227 16.227 0 0 1-5.4-.952l-49.947-17.921a16.04 16.04 0 0 1-9.114-8.261 16.066 16.066 0 0 1-.595-12.289l27.311-76.125a16.057 16.057 0 0 1 8.263-9.114 16.062 16.062 0 0 1 12.289-.594l49.947 17.92a16.068 16.068 0 0 1 9.116 8.262 16.096 16.096 0 0 1 1.519 6.085 16.043 16.043 0 0 1-.926 6.203l-27.311 76.125a16.05 16.05 0 0 1-5.884 7.735 16.047 16.047 0 0 1-9.268 2.926Zm-4.628-3.085a13.785 13.785 0 0 0 10.555-.515 13.773 13.773 0 0 0 7.093-7.833l27.311-76.124a13.8 13.8 0 0 0-3.724-14.881 13.809 13.809 0 0 0-4.624-2.766l-49.948-17.92a13.803 13.803 0 0 0-10.554.514 13.772 13.772 0 0 0-4.328 3.211 13.818 13.818 0 0 0-2.766 4.623l-27.31 76.125a14.01 14.01 0 0 0 .499 10.569 13.738 13.738 0 0 0 3.215 4.327 13.674 13.674 0 0 0 4.634 2.75l49.947 17.92Z"
-            fill="#5624C2"
-         />
-         <path
-            d="M279.064 219.173a9.572 9.572 0 0 0 9.572-9.572c0-5.287-4.285-9.573-9.572-9.573a9.572 9.572 0 0 0-9.573 9.573c0 5.286 4.285 9.572 9.573 9.572Z"
-            fill="#fff"
-         />
-         <path
-            d="M279.063 220.261a11.068 11.068 0 0 1-3.63-.59 10.85 10.85 0 0 1-6.005-5.087 10.859 10.859 0 0 1-.982-7.809 10.84 10.84 0 0 1 19.035-4.063 10.838 10.838 0 0 1-1.112 14.624 10.84 10.84 0 0 1-7.306 2.925Zm0-19.099a8.256 8.256 0 0 0-3.584.816 8.487 8.487 0 0 0-4.355 4.764 8.384 8.384 0 0 0 2.275 9.036 8.385 8.385 0 0 0 9.218 1.366 8.393 8.393 0 0 0 3.61-3.267 8.393 8.393 0 0 0 1.186-4.722 8.39 8.39 0 0 0-5.538-7.494 8.506 8.506 0 0 0-2.812-.499ZM242.999 324.194c12.474 4.672 19.778 3.946 22.456 1.814l-26.132 63.059c-8.665 22.593-39.423 20.823-54.712 1.997l-52.261-62.833 34.841-25.858 38.199 43.868a1.537 1.537 0 0 0 1.383.862c.288 0 .571-.082.815-.232a1.56 1.56 0 0 0 .57-.63c3.221-6.532 11.114-23.362 16.377-34.931 3.357 5.171 6.033 8.212 18.464 12.884Z"
-            fill="#000"
-         />
-         <path
-            d="M214.416 406.717c-11.296 0-23.135-5.626-30.712-14.972l-52.216-62.785a1.219 1.219 0 0 1-.227-.862.987.987 0 0 1 .408-.771l34.841-25.86c.231-.166.513-.245.797-.221.283.026.548.153.746.357l38.198 43.869.136.272a.4.4 0 0 0 .149.164.396.396 0 0 0 .214.063.402.402 0 0 0 .409-.227c3.175-6.532 11.024-23.228 16.331-34.887.09-.182.222-.341.389-.461.168-.118.361-.194.563-.218a1.215 1.215 0 0 1 1.044.545c3.131 4.808 5.534 7.756 17.919 12.428 11.931 4.448 18.965 3.857 21.368 1.953a1.09 1.09 0 0 1 .681-.241c.247 0 .487.086.68.241a1.085 1.085 0 0 1 .362 1.315l-26.13 63.103c-3.539 9.256-10.979 15.243-20.913 16.832a31.574 31.574 0 0 1-5.037.363Zm-80.388-78.302 51.444 61.925c8.212 10.116 21.731 15.652 33.617 13.746 9.119-1.407 15.924-6.895 19.191-15.426l24.996-60.335c-6.579 1.86-17.147-1.724-20.641-3.041-11.07-4.173-14.653-7.077-17.92-11.659-5.262 11.569-12.475 26.902-15.516 33.117a2.678 2.678 0 0 1-2.403 1.497h-.046a2.638 2.638 0 0 1-1.343-.362 2.64 2.64 0 0 1-.97-.998l-37.428-42.963-32.981 24.499Z"
-            fill="#000"
-         />
-         <path
-            d="M277.566 266.08c-3.038-1.043-5.127 1.952-6.125 4.129l-7.53 20.915-12.884-7.078 6.125-20.823a3.09 3.09 0 0 0-1.543-3.902l-1.361-.589a4.31 4.31 0 0 0-3.317-.16 4.349 4.349 0 0 0-1.447.877 4.366 4.366 0 0 0-.998 1.369s-9.8 24.86-11.431 29.624a148.217 148.217 0 0 0-4.311 14.291 42.37 42.37 0 0 1-1.768 4.762c-.818 1.952-1.906 4.355-3.085 6.987 3.311 2.586 8.709 5.444 15.107 7.713 6.395 2.269 11.795 3.403 15.922 3.629l2.814-7.259 18.374-47.951c.589-1.498.499-5.49-2.542-6.534Z"
-            fill="#fff"
-         />
-         <path
-            d="M258.92 328.959h-.044c-4.401-.228-10.344-1.588-16.241-3.675-5.899-2.086-11.885-5.125-15.425-7.893a1.25 1.25 0 0 1-.397-.624 1.236 1.236 0 0 1 .034-.737c1.225-2.722 2.269-5.037 3.085-6.987a35.354 35.354 0 0 0 1.724-4.582c1.543-6.125 3.084-10.934 4.309-14.381 1.633-4.764 11.07-28.672 11.479-29.67v-.09a5.493 5.493 0 0 1 7.303-2.632l1.315.636a4.13 4.13 0 0 1 2.133 5.263l-5.807 19.869 10.934 6.035 7.077-19.691v-.09c1.906-3.993 4.628-5.716 7.53-4.718a5.023 5.023 0 0 1 2.949 2.768 7.17 7.17 0 0 1 .272 5.263l-21.139 55.209a1.14 1.14 0 0 1-.429.537c-.197.13-.427.196-.662.19Zm-29.623-12.839c3.403 2.449 8.438 4.991 14.11 7.031 5.67 2.042 10.661 3.177 14.744 3.493l20.913-54.439a5.314 5.314 0 0 0-.227-3.493 2.728 2.728 0 0 0-.649-.948 2.748 2.748 0 0 0-.984-.593c-2.313-.818-3.947 1.813-4.719 3.447l-7.486 20.869a1.107 1.107 0 0 1-.264.415 1.107 1.107 0 0 1-.415.264 1.165 1.165 0 0 1-.952-.044l-12.886-7.077a1.176 1.176 0 0 1-.543-1.317l6.123-20.823v-.09a1.942 1.942 0 0 0-.02-1.421 1.939 1.939 0 0 0-.976-1.028l-1.317-.635a3.22 3.22 0 0 0-4.265 1.541c-.407 1.044-9.799 24.905-11.431 29.533a182.213 182.213 0 0 0-4.265 14.201 42.111 42.111 0 0 1-1.814 4.945c-.681 1.723-1.589 3.81-2.677 6.169Z"
-            fill="#000"
-         />
-         <path
-            d="M230.977 309.497c-.818 1.951-1.907 4.353-3.086 6.985-2.541-1.996-3.811-3.81-3.356-5.171.453-1.361 2.902-1.996 6.442-1.814Z"
-            fill="#fff"
-         />
-         <path
-            d="M227.891 317.619a1.216 1.216 0 0 1-.681-.227c-3.131-2.449-4.4-4.628-3.72-6.442.681-1.814 3.221-2.768 7.53-2.585a1.082 1.082 0 0 1 .908.543c.094.156.152.331.168.511.014.182-.012.365-.078.533a440.84 440.84 0 0 1-3.084 6.987 1.164 1.164 0 0 1-.772.634l-.271.046Zm1.361-6.987c-2.541.046-3.493.681-3.63 1.09-.136.407.137 1.361 1.86 2.948.68-1.497 1.271-2.858 1.77-4.038Z"
-            fill="#000"
-         />
-         <path
-            d="M265.454 326.009c-.499 1.407-2.948 1.996-6.534 1.815l2.814-7.259c2.766 2.086 4.219 4.039 3.72 5.444Z"
-            fill="#fff"
-         />
-         <path
-            d="M260.147 329.006h-1.271a1.223 1.223 0 0 1-.908-.499 1.173 1.173 0 0 1-.09-1.042l2.768-7.26a1.04 1.04 0 0 1 .769-.679 1.123 1.123 0 0 1 .998.18c3.404 2.542 4.809 4.809 4.129 6.715-.635 1.723-2.768 2.585-6.395 2.585Zm2.086-6.578-1.68 4.311c2.722-.092 3.676-.681 3.857-1.09.182-.407-.227-1.541-2.177-3.221Z"
-            fill="#000"
-         />
-         <path
-            d="m257.605 292.118 7.759-29.667a3.092 3.092 0 0 0-1.59-3.903l-1.314-.589a4.28 4.28 0 0 0-1.642-.411 4.296 4.296 0 0 0-3.123 1.131 4.335 4.335 0 0 0-.998 1.366l-9.799 29.035"
-            fill="#fff"
-         />
-         <path
-            d="M257.651 293.255h-.319a1.176 1.176 0 0 1-.815-1.405l7.712-29.624a.203.203 0 0 0 .036-.064c.008-.024.01-.048.008-.072a1.906 1.906 0 0 0-.952-2.452l-1.314-.633a3.633 3.633 0 0 0-2.496-.092 3.27 3.27 0 0 0-1.77 1.633l-9.753 28.945a1.173 1.173 0 0 1-.222.388 1.124 1.124 0 0 1-.355.275 1.164 1.164 0 0 1-.876.062 1.153 1.153 0 0 1-.663-.575 1.146 1.146 0 0 1-.062-.876l9.799-29.034c0-.047.046-.047.046-.093a5.58 5.58 0 0 1 3.085-2.856 5.668 5.668 0 0 1 4.219.181l1.314.635a4.217 4.217 0 0 1 2.179 5.263l-7.712 29.578a1.185 1.185 0 0 1-1.089.816Z"
-            fill="#000"
-         />
-         <path
-            d="m262.414 296.475 5.217-15.287c.273-.591 3.267-9.119 4.309-13.565 1.181-3.63 1.498-5.127-1.677-6.442-3.175-1.315-4.855 2.133-5.897 4.309l-10.254 27.9"
-            fill="#fff"
-         />
-         <path
-            d="M262.413 297.611a.663.663 0 0 1-.406-.091 1.132 1.132 0 0 1-.682-1.407c.046-.09.409-1.134.908-2.629 4.309-12.658 4.355-12.704 4.4-12.75.046-.044 3.176-8.981 4.219-13.382 0-.013.006-.025.014-.033a.045.045 0 0 1 .032-.012c1.178-3.812 1.134-4.175-1.045-5.081-2.176-.907-2.812.271-4.445 3.72L255.2 293.754a1.142 1.142 0 0 1-.587.62 1.135 1.135 0 0 1-1.544-1.389l10.252-27.856c0-.046 0-.09.046-.09.906-1.997 3.085-6.669 7.348-4.901 4.265 1.77 3.404 4.446 2.359 7.803-1.088 4.355-3.946 12.657-4.399 13.655-.455.998-2.948 8.711-4.311 12.657-.862 2.632-.862 2.632-.906 2.632a1.138 1.138 0 0 1-1.045.726ZM223.264 308.997l-16.83 34.797-46.91-62.605"
-            fill="#000"
-         />
-         <path
-            d="M206.432 344.926a1.036 1.036 0 0 1-.907-.455l-46.909-62.604a1.134 1.134 0 1 1 1.814-1.361l45.774 61.153 16.061-33.163a1.138 1.138 0 0 1 1.497-.499 1.09 1.09 0 0 1 .543 1.497l-16.876 34.797a1.046 1.046 0 0 1-.906.635h-.091Z"
-            fill="#fff"
-         />
-         <path
-            d="M118.378 313.944c.816.59 67.594 82.612 67.594 82.612s13.928 12.748 27.357 11.93c13.429-.816 18.736-5.897 24.769-11.93"
-            fill="#000"
-         />
-         <path
-            d="M212.014 409.664c-5.446 0-11.479-1.996-18.011-5.853a59.6 59.6 0 0 1-8.801-6.441l-.136-.137c-22.638-27.81-65.782-80.705-67.46-82.475a1.135 1.135 0 0 1-.136-1.497c.18-.241.449-.399.747-.443.298-.042.6.036.841.216.816.589 42.734 52.034 67.731 82.702.999.908 14.109 12.385 26.449 11.615 13.293-.818 18.282-5.808 24.09-11.615a1.135 1.135 0 0 1 1.588 1.587c-5.853 5.899-11.387 11.433-25.541 12.295l-1.361.046ZM398.603 232.059l-9.072-45.548-25.677 5.081 8.935 44.822c-11.522 3.175-1.994 9.617 14.247 6.396 16.241-3.221 19.642-11.25 11.567-10.751Z"
-            fill="#fff"
-         />
-         <path
-            d="M378.552 244.85c-6.759 0-11.342-1.86-11.932-4.401-.272-1.135.046-3.268 4.855-4.809l-8.711-43.824a1.103 1.103 0 0 1 .15-.842 1.096 1.096 0 0 1 .712-.474l25.677-5.126c.301-.048.607.016.862.182.241.173.403.434.455.725l8.891 44.595c3.766.09 4.355 2.133 4.402 2.814.589 4.399-8.168 8.528-16.65 10.252-2.868.585-5.785.89-8.711.908Zm-13.337-52.353 8.709 43.732a1.23 1.23 0 0 1-.172.804 1.229 1.229 0 0 1-.644.511c-3.447.908-4.355 1.997-4.265 2.405.363 1.452 7.077 3.903 17.965 1.77 10.889-2.132 15.062-6.398 14.882-7.758-.092-.453-1.135-.906-3.041-.77a1.226 1.226 0 0 1-1.178-.908l-8.847-44.459-23.409 4.673Z"
-            fill="#000"
-         />
-         <path
-            d="M387.443 208.92c-4.626 5.58-10.661 9.346-17.646 10.389-18.057 2.586-35.433-14.29-38.834-37.744-3.403-23.455 8.528-44.55 26.583-47.181 18.057-2.631 35.433 14.29 38.834 37.744 1.543 10.752-.181 21.368-4.173 29.579"
-            fill="#fff"
-         />
-         <path
-            d="M366.076 220.715c-7.805 0-15.471-3.357-22.003-9.708-7.486-7.304-12.567-17.693-14.245-29.307-1.679-11.613.273-22.955 5.398-32.073 5.127-9.119 13.066-15.062 22.139-16.378 9.074-1.315 18.282 2.042 25.904 9.436 7.623 7.395 12.567 17.693 14.199 29.262 1.543 10.57 0 21.594-4.263 30.259a1.076 1.076 0 0 1-1.071.661 1.085 1.085 0 0 1-.427-.117 1.136 1.136 0 0 1-.545-1.497c4.039-8.302 5.536-18.872 4.083-28.944-1.588-11.114-6.442-21.049-13.563-27.99-7.124-6.941-15.606-10.026-23.954-8.801-8.348 1.225-15.698 6.623-20.506 15.197-4.808 8.574-6.714 19.598-5.127 30.668 1.59 11.069 6.398 21.05 13.565 27.991 7.168 6.941 15.606 10.026 23.954 8.801 6.488-.953 12.34-4.401 16.966-9.981a1.14 1.14 0 0 1 1.512-.043 1.132 1.132 0 0 1 .212 1.495c-4.945 6.034-11.296 9.754-18.374 10.752a22.118 22.118 0 0 1-3.854.317Z"
-            fill="#000"
-         />
-         <path
-            d="M355.28 192.635c-1.816-.045-10.708-.499-11.523-4.537-.183-1.088.543-2.177 1.723-3.765 2.267-3.221 6.124-8.619 4.127-16.876a1.149 1.149 0 0 1 .151-.859 1.132 1.132 0 0 1 1.156-.513 1.149 1.149 0 0 1 .738.466c.086.123.146.261.178.407 2.225 9.209-2.177 15.379-4.536 18.691a11 11 0 0 0-1.27 2.041c.455 1.588 5.626 2.586 9.3 2.722.148-.007.298.018.437.073.138.055.262.139.364.247a1.029 1.029 0 0 1 .287.814 1.135 1.135 0 0 1-1.132 1.089ZM402.323 165.774c2.722 16.196-8.166 31.212-8.166 31.212s6.035-21.004-1.587-28.036c-3.766-3.402-10.98-3.04-16.151-7.576-9.3-8.121-12.793-19.371-7.756-25.088 5.034-5.716 16.557-3.765 25.857 4.401s6.578 17.602 7.803 25.087Z"
-            fill="#000"
-         />
-         <path
-            d="m394.157 198.123-.499-.091a1.185 1.185 0 0 1-.589-1.361 69.398 69.398 0 0 0 2.042-11.205c.453-5.308.271-12.34-3.313-15.651-1.587-1.452-3.992-2.178-6.532-2.949-2.541-.772-6.624-2.042-9.572-4.628-9.753-8.529-13.293-20.505-7.893-26.675 5.397-6.17 17.738-4.264 27.491 4.264 7.35 6.397 7.53 13.61 7.758 19.916-.01 1.958.127 3.914.407 5.852 2.722 16.468-7.939 31.439-8.392 32.074a1.044 1.044 0 0 1-.908.454Zm-17.239-64.012c-3.084 0-5.716.953-7.44 2.949-4.582 5.217-1.134 15.742 7.713 23.454 2.585 2.269 5.716 3.221 8.755 4.174 3.04.953 5.445 1.679 7.396 3.448 5.125 4.718 4.581 14.971 3.447 21.957 2.812-5.761 5.989-14.789 4.446-24.135a38.166 38.166 0 0 1-.453-6.124c-.227-6.17-.409-12.521-6.988-18.283-5.489-4.808-11.795-7.44-16.876-7.44Z"
-            fill="#000"
-         />
-         <path
-            d="M389.531 169.633c1.814-5.489 6.213-9.073 10.569-8.257 4.355.817 7.849 7.123 6.624 13.792-1.225 6.669-6.261 11.341-11.16 10.434a7.815 7.815 0 0 1-5.037-3.629"
-            fill="#fff"
-         />
-         <path
-            d="m396.788 186.827-1.451-.136c-2.315-.408-4.355-1.906-5.762-4.128a1.087 1.087 0 0 1 .362-1.543 1.184 1.184 0 0 1 1.588.317 6.37 6.37 0 0 0 4.219 3.131c2.042.408 4.129-.363 5.989-2.042a14.604 14.604 0 0 0 2.948-15.832c-1.088-2.269-2.766-3.72-4.762-4.083-3.72-.726-7.622 2.449-9.3 7.44-.09.288-.293.53-.559.67a1.145 1.145 0 0 1-1.542-.481 1.144 1.144 0 0 1-.078-.87c2.043-6.124 7.033-9.89 11.886-8.982 2.724.499 4.991 2.404 6.398 5.353s1.723 6.26 1.088 9.754a15.893 15.893 0 0 1-4.536 8.665c-1.952 1.814-4.219 2.767-6.488 2.767Z"
-            fill="#000"
-         />
-         <path
-            d="M389.529 174.713c-1.541-5.353-7.666-10.933-12.338-13.973l17.056 1.361c-2.267 3.493-3.764 8.801-4.718 12.612Z"
-            fill="#000"
-         />
-         <path
-            d="M389.529 175.847a1.18 1.18 0 0 1-1.089-.816c-1.361-4.764-6.985-10.117-11.885-13.338a1.098 1.098 0 0 1-.465-.576 1.079 1.079 0 0 1 .012-.739 1.126 1.126 0 0 1 1.178-.772l17.059 1.361a1.176 1.176 0 0 1 .952.635 1.136 1.136 0 0 1-.09 1.134c-1.68 2.586-3.221 6.715-4.538 12.249a1.179 1.179 0 0 1-1.088.862h-.046Zm-8.437-13.655c3.129 2.495 6.306 5.716 8.166 9.073a42.884 42.884 0 0 1 3.129-8.166l-11.295-.907Z"
-            fill="#000"
-         />
-         <path
-            d="M399.511 187.416c-3.674-.227-6.941-3.72-8.438-6.533l-4.173-19.507-6.125 1.225 4.627 23.182c.59 2.949-1.36 6.215-4.309 6.215-17.966.045-44.369 5.126-44.05 7.848 1.405 12.022 6.94 25.087 19.235 28.082 4.127.998 32.39-.136 40.92-12.294 6.123-8.71 2.313-28.218 2.313-28.218Z"
-            fill="#000"
-         />
-         <path
-            d="M360.405 229.334a23.074 23.074 0 0 1-4.4-.318c-10.797-2.586-18.283-13.428-20.097-29.034a1.797 1.797 0 0 1 .102-.734c.086-.234.223-.448.397-.627 2.948-3.312 28.399-7.712 44.684-7.757.445-.007.88-.118 1.275-.322.393-.205.733-.499.994-.858a4.426 4.426 0 0 0 .864-1.736 4.437 4.437 0 0 0 .044-1.939l-4.628-23.182a1.131 1.131 0 0 1 .182-.862c.177-.244.433-.421.726-.499l6.123-1.225c.295-.052.599.013.846.183.246.17.415.431.471.725l4.173 19.326c1.497 2.767 4.492 5.625 7.396 5.807a1.134 1.134 0 0 1 1.042.907c.182.817 3.856 20.097-2.449 29.08-4.764 6.805-15.063 10.071-22.819 11.568a83.108 83.108 0 0 1-14.926 1.497Zm-22.184-29.306c.681 5.761 4.039 23.318 18.283 26.765 4.083.998 31.71-.317 39.786-11.795 5.081-7.258 2.858-22.819 2.223-26.539-3.63-.68-6.805-3.947-8.439-7.077a.626.626 0 0 1-.09-.272l-3.946-18.419-3.949.771 4.402 22.094a6.813 6.813 0 0 1-.062 2.94 6.85 6.85 0 0 1-1.299 2.64 5.227 5.227 0 0 1-4.039 1.996c-7.71 0-17.918.998-27.355 2.677-9.436 1.678-14.607 3.447-15.515 4.219Z"
-            fill="#000"
-         />
-         <path
-            d="m351.603 201.707 16.378-3.267s-1.634 7.713-6.759 8.892c-5.127 1.18-9.619-5.625-9.619-5.625Z"
-            fill="#fff"
-         />
-         <path
-            d="M360.042 208.602c-5.125 0-9.164-5.989-9.344-6.261a1.17 1.17 0 0 1-.137-1.089 1.232 1.232 0 0 1 .816-.68l16.377-3.221c.187-.039.381-.03.564.025.182.056.346.156.481.293.124.13.214.29.26.464.048.174.052.358.01.534-.044.362-1.814 8.438-7.576 9.753-.473.125-.96.186-1.451.182Zm-6.396-6.17c1.497 1.769 4.355 4.446 7.35 3.765 2.992-.68 4.626-4.128 5.351-6.26l-12.701 2.495ZM360.485 127.163l-29.185 5.831c-8.254 1.65-13.609 9.676-11.962 17.928 1.65 8.251 9.679 13.603 17.933 11.954l29.185-5.832c8.253-1.649 13.611-9.676 11.961-17.927-1.649-8.252-9.678-13.604-17.932-11.954Z"
-            fill="#000"
-         />
-         <path
-            d="M334.184 164.688a16.334 16.334 0 0 1-3.177-32.392l29.171-5.807a16.353 16.353 0 0 1 6.396-.024 16.332 16.332 0 0 1 10.44 6.953 16.311 16.311 0 0 1 2.446 5.91c.42 2.109.422 4.281.004 6.391a16.34 16.34 0 0 1-12.888 12.844l-29.171 5.807c-1.06.218-2.14.324-3.221.318Zm29.261-36.248c-.944.001-1.886.092-2.814.272l-29.169 5.807a14.08 14.08 0 0 0-5.083 2.097 14.07 14.07 0 0 0-5.986 14.462 14.06 14.06 0 0 0 16.557 11.069l29.217-5.807a14.064 14.064 0 0 0-2.722-27.9Z"
-            fill="#000"
-         />
-         <path
-            d="M393.271 141.774c2.183-7.009-3.106-14.889-11.813-17.601-8.707-2.712-17.536.771-19.719 7.78-2.182 7.009 3.105 14.889 11.814 17.601 8.707 2.713 17.536-.771 19.718-7.78Z"
-            fill="#000"
-         />
-         <path
-            d="M379.822 151.895a20.454 20.454 0 0 1-6.262-.998 19.51 19.51 0 0 1-10.57-7.621c-2.45-3.63-3.129-7.667-1.995-11.387 2.404-7.622 11.886-11.433 21.14-8.529 9.3 2.903 14.926 11.432 12.567 19.053-1.816 5.853-7.895 9.482-14.88 9.482Zm-3.993-27.22c-5.942 0-11.114 2.995-12.657 7.894-.952 3.04-.363 6.397 1.678 9.436a17.603 17.603 0 0 0 9.392 6.714c8.075 2.541 16.285-.589 18.281-6.986 1.997-6.397-2.948-13.655-11.023-16.196a19.162 19.162 0 0 0-5.671-.862Z"
-            fill="#000"
-         />
-         <path
-            d="M354.764 148.045c10.785-2.155 18.363-9.738 16.924-16.937-1.439-7.199-11.348-11.288-22.133-9.132-10.788 2.155-18.364 9.738-16.927 16.937 1.439 7.198 11.349 11.287 22.136 9.132Z"
-            fill="#000"
-         />
-         <path
-            d="M349.018 150.17c-8.938 0-16.331-4.174-17.602-10.616-1.588-7.803 6.442-15.969 17.828-18.237 5.444-1.089 10.844-.68 15.243 1.134 4.401 1.815 7.486 5.036 8.258 8.892 1.541 7.803-6.488 15.969-17.875 18.237a29.12 29.12 0 0 1-5.852.59Zm6.169-27.22a28.993 28.993 0 0 0-5.488.59c-10.162 1.996-17.375 8.982-16.06 15.561 1.314 6.578 10.614 10.298 20.778 8.256 10.162-2.041 17.374-9.028 16.059-15.56-.589-3.131-3.039-5.671-6.851-7.259a22.002 22.002 0 0 0-8.438-1.588ZM398.149 181.881a1.13 1.13 0 0 1-1.066-.825 1.135 1.135 0 0 1 .477-1.262c3.311-2.177 4.4-5.761 3.402-10.978a1.081 1.081 0 0 1 .166-.856 1.078 1.078 0 0 1 .742-.46 1.134 1.134 0 0 1 1.314.862c1.225 6.124-.27 10.616-4.399 13.338a1.498 1.498 0 0 1-.636.181ZM362.132 173.786c1.355-.249 2.084-2.468 1.627-4.957-.457-2.489-1.926-4.305-3.281-4.056-1.357.248-2.084 2.468-1.627 4.957.457 2.489 1.926 4.305 3.281 4.056ZM340.738 178.25c1.357-.249 2.085-2.468 1.628-4.957-.457-2.489-1.926-4.305-3.281-4.056-1.355.248-2.085 2.468-1.628 4.957.457 2.489 1.926 4.305 3.281 4.056Z"
-            fill="#000"
-         />
-         <path
-            d="M340.853 135.473a1.116 1.116 0 0 1-.474-.084 1.094 1.094 0 0 1-.388-.279c-2.043-2.268-7.124-3.947-14.608-1.043a1.137 1.137 0 0 1-1.497-.636 1.131 1.131 0 0 1 .679-1.451c11.024-4.31 16.197.59 17.149 1.633.197.224.297.515.279.812a1.128 1.128 0 0 1-.369.776 1.229 1.229 0 0 1-.771.272ZM367.527 160.65a1.13 1.13 0 0 1-1.134-.907 1.135 1.135 0 0 1 .908-1.316c3.856-.726 4.309-2.495 4.309-3.765a1.136 1.136 0 0 1 1.135-1.135c.3 0 .589.12.801.333.213.212.333.501.333.802 0 2.223-1.044 5.081-6.171 5.988h-.181Z"
-            fill="#fff"
-         />
-         <path
-            d="M184.566 268.213a3.478 3.478 0 0 1-1.94-.589 3.485 3.485 0 0 1-1.287-1.567 3.505 3.505 0 0 1-.2-2.019 3.49 3.49 0 0 1 4.764-2.545 3.495 3.495 0 0 1 2.157 3.227 3.46 3.46 0 0 1-1.01 2.485 3.477 3.477 0 0 1-1.141.756 3.462 3.462 0 0 1-1.343.252Zm0-4.718a1.226 1.226 0 1 0 .001 2.452 1.226 1.226 0 0 0-.001-2.452ZM187.424 284.952a3.99 3.99 0 0 1-3.916-4.77 3.993 3.993 0 0 1 7.909.777 3.941 3.941 0 0 1-.293 1.536 3.949 3.949 0 0 1-3.7 2.457Zm0-5.716a1.727 1.727 0 0 0-1.218 2.944 1.718 1.718 0 0 0 1.878.373 1.726 1.726 0 0 0 .559-2.812 1.725 1.725 0 0 0-1.219-.505Z"
-            fill="#000"
-         />
-         {/* holding swatch border and fill */}
-         <path d="m272.929 382.35-16.091 17.621 67.367 61.515 16.091-17.622-67.367-61.514Z" fill="#fff" />
-         <path
-            d="M324.202 462.88c-.28 0-.553-.096-.771-.273l-67.323-61.562a1.132 1.132 0 0 1-.09-1.588l16.104-17.602a1.144 1.144 0 0 1 .816-.362c.299-.02.591.078.818.272l67.323 61.563c.116.09.21.208.272.342a.93.93 0 0 1 .09.427 1.26 1.26 0 0 1-.272.818l-16.105 17.602a1.094 1.094 0 0 1-.862.363Zm-65.735-62.741 65.691 60.019 14.562-15.924-65.69-60.019-14.563 15.924Z"
-            fill="#000"
-         />
-         {/* holding swatch light shade */}
-         <path d="m284.125 392.159-16.365 17.923 12.595 11.502 16.367-17.922-12.597-11.503Z" fill="#5624C2" />
-         <path
-            d="m280.47 423.409-14.291-13.02 17.919-19.598 14.246 13.066-17.874 19.552Zm-11.07-13.156 10.888 9.981 14.88-16.241-10.934-9.981-14.834 16.241Z"
-            fill="#5624C2"
-         />
-         {/* holding swatch dark shade */}
-         <path d="m302.247 409.355-15.785 17.288 12.597 11.502 15.784-17.287-12.596-11.503Z" fill="#000" />
-         <path
-            d="M299.025 439.606a1.042 1.042 0 0 1-.772-.319l-12.565-11.476a1.15 1.15 0 0 1-.363-.818c-.02-.297.077-.591.271-.816l15.834-17.285a1.132 1.132 0 0 1 1.588-.09l12.611 11.523c.2.212.328.481.362.771a1.043 1.043 0 0 1-.316.816l-15.788 17.329a1.231 1.231 0 0 1-.862.365Zm-10.978-12.703 10.932 9.979 14.246-15.65-10.889-9.981-14.289 15.652Z"
-            fill="#000"
-         />
-         <path d="m320.543 425.38-16.368 17.922 12.597 11.503 16.366-17.922-12.595-11.503Z" fill="#829EFF" />
-         <path
-            d="m316.853 456.712-14.245-13.066 17.921-19.598 14.244 13.066-17.92 19.598Zm-11.068-13.202 10.931 9.981 14.836-16.241-10.887-9.981-14.88 16.241Z"
-            fill="#829EFF"
-         />
-         <path d="M270.671 398.232a2.496 2.496 0 1 0 0-4.993 2.496 2.496 0 0 0 0 4.993Z" fill="#fff" />
-
-         {/* holding swatch dot */}
-         <path
-            d="M270.672 399.366a3.639 3.639 0 0 1-2.568-1.062 3.634 3.634 0 0 1 0-5.133 3.626 3.626 0 0 1 2.568-1.062c.479-.006.956.082 1.403.262a3.633 3.633 0 0 1 2.271 3.365 3.718 3.718 0 0 1-1.089 2.586c-.34.339-.743.605-1.188.784-.443.18-.918.268-1.397.26Zm0-4.99a1.37 1.37 0 0 0-.964.398 1.364 1.364 0 0 0-.397.962 1.24 1.24 0 0 0 .098.524c.07.164.177.31.309.43a1.312 1.312 0 0 0 .954.407 1.405 1.405 0 0 0 1.405-1.361c0-.182-.036-.362-.106-.531a1.368 1.368 0 0 0-1.299-.829Z"
-            fill="#000"
-         />
-         <path
-            d="m435.169 279.465-49.767 15.968s22.593 61.108 12.748 68.504c-5.807 4.399-59.067 36.608-70.317 42.417a38.315 38.315 0 0 1-4.672 2.04c-3.72 1.135-22.775-.725-22.775-.725l7.213 6.351c-4.672 1.904-20.28 8.483-20.28 8.483-2.13 1.134-7.484 4.175-5.714 7.622 1.768 3.447 6.486 2.177 7.939 1.497l13.882-3.902 7.757 12.477-6.26 2.449c-2.132 1.134-5.036 3.357-3.812 6.352 1.225 2.994 5.173 2.858 6.624 2.176l28.716-12.883s83.566-35.068 100.486-48.407c21.595-17.329-1.768-110.419-1.768-110.419Z"
-            fill="#fff"
-         />
-         <path
-            d="M305.602 452.718a6.982 6.982 0 0 1-2.675-.5 4.852 4.852 0 0 1-2.858-2.812c-1.181-2.858.406-5.672 4.309-7.758h.136l4.991-1.95-6.578-10.617-12.976 3.676c-1.95.862-7.258 2.086-9.344-2.043-2.089-4.127 2.267-7.121 6.169-9.163h.09l18.6-7.803-5.806-5.127a1.118 1.118 0 0 1-.375-.608 1.115 1.115 0 0 1 .062-.711c.096-.225.263-.411.475-.533.211-.123.457-.173.7-.143 5.217.544 19.417 1.678 22.32.77a39.682 39.682 0 0 0 4.491-1.95c11.932-6.169 64.601-38.152 70.137-42.327 2.087-1.541 4.399-8.256-3.087-35.839-4.355-15.969-9.979-31.211-10.069-31.394a1.131 1.131 0 0 1 .32-1.23c.119-.101.255-.177.405-.221l49.767-15.968c.147-.056.307-.076.463-.06.157.016.309.068.443.15.263.148.457.393.546.679.952 3.857 23.318 94.045 1.36 111.603-7.077 5.624-26.222 15.515-56.798 29.397-22.775 10.298-43.734 19.099-43.96 19.189l-28.626 12.884a6.32 6.32 0 0 1-2.632.409Zm-.226-9.028c-1.634.862-4.129 2.632-3.221 4.855a2.957 2.957 0 0 0 1.634 1.587 5.348 5.348 0 0 0 3.493.044l28.67-12.928c.862-.362 83.656-35.159 100.259-48.27 19.508-15.651.726-97.265-1.86-108.062l-47.498 15.197c1.361 3.856 5.989 16.922 9.709 30.487 6.079 22.32 6.805 34.795 2.269 38.243-6.215 4.674-59.566 36.883-70.456 42.553a39.897 39.897 0 0 1-4.852 2.087c-2.994.908-13.974.136-19.781-.363l4.582 4.039a1.035 1.035 0 0 1 .363 1.042 1.04 1.04 0 0 1-.679.862c-4.492 1.815-19.282 8.075-20.234 8.484-2.948 1.587-6.261 3.902-5.127 6.079 1.543 3.04 6.398.998 6.442.998l.182-.09 13.836-3.903a1.09 1.09 0 0 1 1.271.5l7.757 12.476a1.165 1.165 0 0 1 .136.952 1.315 1.315 0 0 1-.68.726l-6.215 2.405Z"
-            fill="#000"
-         />
-         <path
-            d="m311.817 433.892-16.241 4.762c-6.033 2.043-7.849-1.132-7.849-1.132a4.447 4.447 0 0 1 1.816-5.899l18.69-8.302"
-            fill="#fff"
-         />
-         <path
-            d="M291.903 440.47c-3.72 0-5.083-2.269-5.127-2.405a5.208 5.208 0 0 1-.409-4.173 5.348 5.348 0 0 1 2.677-3.267h.045l18.692-8.348a1.08 1.08 0 0 1 .878-.034c.14.054.268.138.374.247.107.108.191.236.243.378a1.134 1.134 0 0 1-.543 1.496l-18.692 8.303a3.086 3.086 0 0 0-1.496 1.86 3.27 3.27 0 0 0 .181 2.45c.136.18 1.679 2.222 6.488.589l16.287-4.764c.142-.04.292-.054.439-.038a1.128 1.128 0 0 1 .966.809 1.1 1.1 0 0 1 .038.441 1.118 1.118 0 0 1-.417.762 1.152 1.152 0 0 1-.393.204l-16.241 4.763a12.264 12.264 0 0 1-3.99.727Z"
-            fill="#000"
-         />
-         <path
-            d="M316.447 438.427c0 .046-11.842 5.217-16.197 6.532-1.86.635-5.081 1.77-6.532-1.361-1.451-3.129 1.86-4.989 3.992-6.123l15.243-7.167"
-            fill="#fff"
-         />
-         <path
-            d="M296.938 446.82a4.404 4.404 0 0 1-4.265-2.722c-1.996-4.173 2.586-6.578 4.492-7.62h.046l15.243-7.168a1.118 1.118 0 0 1 .92-.122 1.121 1.121 0 0 1 .703.607 1.124 1.124 0 0 1 .014.928 1.166 1.166 0 0 1-.274.387 1.147 1.147 0 0 1-.411.241l-15.197 7.167c-3.766 1.996-4.129 3.221-3.493 4.582.635 1.361 2.539 1.68 5.127.772h.09c4.263-1.271 15.968-6.396 16.058-6.486a1.148 1.148 0 0 1 1.522.565 1.152 1.152 0 0 1-.568 1.521c-.136.044-11.887 5.217-16.377 6.578a11.74 11.74 0 0 1-3.63.77Z"
-            fill="#000"
-         />
-         <path
-            d="m443.199 267.079 7.893 36.474c1.27 6.398-12.974 14.61-31.847 18.374-18.872 3.766-35.159 1.633-36.428-4.762l-10.072-33.709"
-            fill="#829EFF"
-         />
-         <path
-            d="M401.779 324.878c-10.571 0-18.963-2.315-20.053-7.442l-10.07-33.614a1.056 1.056 0 0 1-.054-.445 1.11 1.11 0 0 1 .13-.429 1.084 1.084 0 0 1 .694-.533 1.133 1.133 0 0 1 1.407.725l10.071 33.663a.178.178 0 0 0 .044.136c.499 2.539 4.356 4.445 10.663 5.261 6.306.818 15.424.409 24.451-1.407 9.029-1.814 17.195-4.626 23.092-8.029 5.898-3.401 8.348-6.442 7.848-8.981l-7.892-36.431a1.13 1.13 0 0 1-.015-.445c.027-.146.081-.288.161-.415.08-.126.184-.236.306-.32.125-.086.263-.149.409-.181a1.185 1.185 0 0 1 1.361.862l7.893 36.475c1.588 7.939-15.425 16.241-32.754 19.69a90.276 90.276 0 0 1-17.692 1.86ZM365.648 260.809c2.984-1.175 5.069-2.979 4.656-4.027-.415-1.05-3.169-.948-6.151.228-2.985 1.175-5.069 2.979-4.656 4.027.412 1.05 3.166.948 6.151-.228Z"
-            fill="#fff"
-         />
-         {/* middle swatch fill AND shirt spots */}
-         <path
-            d="M405.836 300.709c.471-1.024-1.51-2.94-4.422-4.281-2.914-1.341-5.658-1.595-6.129-.571-.471 1.024 1.509 2.94 4.421 4.281 2.915 1.339 5.658 1.595 6.13.571ZM411.193 261.402c2.58-1.902 4.131-4.181 3.462-5.089-.67-.906-3.303-.1-5.885 1.804-2.581 1.902-4.133 4.181-3.463 5.089.669.908 3.305.1 5.886-1.804ZM437.164 301.597c1.603-2.778 2.112-5.486 1.136-6.051-.976-.563-3.068 1.23-4.672 4.008-1.603 2.778-2.112 5.486-1.134 6.049.976.566 3.066-1.23 4.67-4.006ZM348.455 303.866c.936-.63.244-3.297-1.545-5.959-1.788-2.662-3.997-4.309-4.933-3.682-.936.629-.245 3.297 1.545 5.959 1.788 2.662 3.997 4.309 4.933 3.682ZM366.275 329.321c1.956-2.542 2.818-5.159 1.926-5.847-.894-.687-3.205.814-5.161 3.355-1.957 2.542-2.818 5.16-1.927 5.847.894.688 3.205-.814 5.162-3.355ZM344.977 352.075c.381-1.062-1.759-2.8-4.778-3.88-3.02-1.08-5.774-1.097-6.155-.036-.381 1.062 1.76 2.8 4.778 3.88 3.021 1.08 5.777 1.096 6.155.036ZM386.194 357.753c2.778-1.603 4.572-3.694 4.009-4.67-.563-.978-3.273-.469-6.049 1.135-2.778 1.603-4.574 3.696-4.009 4.672.563.976 3.271.467 6.049-1.137ZM342.416 389.546c2.267-2.267 3.459-4.752 2.662-5.55-.798-.796-3.281.395-5.55 2.664-2.267 2.267-3.46 4.752-2.662 5.55.796.795 3.281-.395 5.55-2.664ZM456.229 345.468c1.05-.413.946-3.167-.228-6.151-1.175-2.984-2.979-5.069-4.029-4.654-1.048.413-.946 3.167.23 6.149 1.175 2.985 2.979 5.069 4.027 4.656ZM455.478 385.965c2.267-2.269 3.46-4.753 2.662-5.55-.798-.798-3.281.395-5.55 2.661-2.267 2.269-3.459 4.753-2.662 5.55.798.798 3.282-.395 5.55-2.661ZM440.47 411.751c.469-1.026-1.517-2.937-4.435-4.265-2.919-1.331-5.665-1.578-6.132-.552-.467 1.027 1.52 2.937 4.438 4.266 2.918 1.33 5.662 1.577 6.129.551ZM458.547 441.313c1.752-2.688 2.405-5.366 1.461-5.981-.946-.615-3.13 1.064-4.882 3.75-1.752 2.688-2.405 5.364-1.461 5.979.946.617 3.131-1.062 4.882-3.748ZM412.543 438.09c.976-.564.467-3.273-1.136-6.049-1.604-2.778-3.694-4.572-4.67-4.009-.978.563-.469 3.273 1.134 6.049 1.604 2.778 3.696 4.572 4.672 4.009ZM378.804 437.207c.192-1.113-2.215-2.455-5.376-3.001-3.158-.545-5.876-.088-6.069 1.023-.192 1.112 2.215 2.455 5.376 3 3.159.547 5.877.088 6.069-1.022ZM275.377 499.662l-23.186-16.569 89.889-41.802 17.234 12.284-83.937 46.087Z"
-            fill="#fff"
-         />
-         {/* border  */}
-         <path
-            d="M275.377 500.854a1.152 1.152 0 0 1-.667-.238l-23.186-16.523a1.144 1.144 0 0 1 .191-2.046l89.841-41.849a1.322 1.322 0 0 1 1.189.143l17.284 12.282a1.1 1.1 0 0 1 .475 1.048 1.094 1.094 0 0 1-.618.953l-83.937 46.087a1.162 1.162 0 0 1-.572.143Zm-20.805-17.522 20.9 14.903 81.604-44.753-15.139-10.808-87.365 40.658Z"
-            fill="#210E10"
-         />
-         {/* neutral shade */}
-         <path d="m329.558 469.668-17.997-14.188 15.665-7.522 17.997 13.094-15.665 8.616Z" fill="#fff" />
-         <path
-            d="M329.558 470.857a1.42 1.42 0 0 1-.761-.237L310.8 456.432a1.14 1.14 0 0 1-.429-1.048 1.198 1.198 0 0 1 .666-.953l15.665-7.522a1.294 1.294 0 0 1 1.237.095l17.997 13.094c.168.114.303.271.387.457.084.185.116.389.088.591a1.059 1.059 0 0 1-.618.953l-15.665 8.615a1.247 1.247 0 0 1-.57.143Zm-15.759-15.092 15.854 12.473 13.378-7.33-15.902-11.569-13.33 6.426Z"
-            fill="#210E10"
-         />
-         {/* light shade */}
-         <path d="m306.847 482.14-18.805-15.71 14.807-7.047 20.187 14.142-16.189 8.615Z" fill="#B682FF" />
-         <path
-            d="m306.704 483.569-20.852-17.425 17.139-8.14 22.282 15.662-18.569 9.903Zm-16.472-16.853 16.758 13.997 13.759-7.331-17.995-12.665-12.522 5.999Z"
-            fill="#B682FF"
-         />
-         {/* mid shade */}
-         <path d="m284.137 495.044-21.472-17.045 17.091-7.997 20.092 16.424-15.711 8.618Z" fill="#5624C2" />
-         <path
-            d="m284.041 496.472-23.613-18.759 19.472-9.093 22.092 17.997-17.951 9.855Zm-19.139-18.235 19.379 15.379 13.473-7.379-18.187-14.855-14.665 6.855Z"
-            fill="#5624C2"
-         />
-         {/* middle swatch dot */}
-         <path d="M344.413 453.24a2.618 2.618 0 1 0 .001-5.237 2.618 2.618 0 0 0-.001 5.237Z" fill="#fff" />
-         <path
-            d="M344.459 454.433a5.124 5.124 0 0 1-1.236-.19 3.867 3.867 0 0 1-2.495-1.872 3.838 3.838 0 0 1-.475-1.52 3.84 3.84 0 0 1 .986-2.96c.36-.398.801-.715 1.292-.932a3.83 3.83 0 0 1 1.56-.33c.537 0 1.069.111 1.559.33.477.16.917.414 1.294.745.378.332.687.736.908 1.186.221.452.352.942.381 1.445a3.784 3.784 0 0 1-.202 1.479 3.712 3.712 0 0 1-1.903 2.238 4.068 4.068 0 0 1-1.669.381Zm-.046-5.284c-.214.002-.427.05-.62.143a1.414 1.414 0 0 0-.713.856 1.46 1.46 0 0 0 .048 1.096c.086.17.208.322.355.444.148.122.318.214.501.269a1.395 1.395 0 0 0 1.102-.057 1.382 1.382 0 0 0 .707-.848c.059-.176.084-.366.072-.553a1.45 1.45 0 0 0-.491-.98 1.454 1.454 0 0 0-.485-.276l-.476-.094Z"
-            fill="#210E10"
-         />
-
-         {/* RIGHT SWATCH GALLERY */}
-         {/* right swacth fill */}
-         <path d="m412.018 494.236 40.517-10.427-66.179-48.611-24.375 5.475 50.037 53.563Z" fill="#fff" />
-         <path
-            d="M412.018 495.425a1.195 1.195 0 0 1-.467-.101 1.231 1.231 0 0 1-.389-.28l-50.086-53.561a1.103 1.103 0 0 1-.284-.538 1.08 1.08 0 0 1 .047-.606 1.08 1.08 0 0 1 .309-.528c.151-.145.341-.242.547-.28l24.423-5.523c.166-.034.339-.03.503.012.166.04.32.118.45.225l66.179 48.611c.196.125.349.305.442.518a1.155 1.155 0 0 1-.261 1.275 1.167 1.167 0 0 1-.61.303l-40.517 10.426-.286.047Zm-47.753-54.038 48.134 51.516 37.423-9.617-63.704-46.802-21.853 4.903Z"
-            fill="#210E10"
-         />
-         {/* middle shade swatch */}
-         <path d="m377.643 457.479 28.376-7.857-6.712-5.235-28.376 6.283 6.712 6.809Z" fill="#5624C2" />
-         <path
-            d="m377.311 458.812-8.761-8.855 31.043-6.857 8.999 7-31.281 8.712Zm-4.047-7.427 4.76 4.76 25.379-7.046-4.382-3.427-25.757 5.713Z"
-            fill="#5624C2"
-         />
-         {/* dark shade swatch */}
-         <path d="m390.213 470.905 33.279-8.475-11.474-8.38-29.756 7.999 7.951 8.856Z" fill="#210E10" />
-         <path
-            d="M390.214 472.097a1.287 1.287 0 0 1-.904-.381l-7.951-8.856a1.142 1.142 0 0 1-.238-1.144c.05-.193.153-.37.294-.513.143-.141.32-.244.516-.295l29.755-8.047c.36-.059.732.008 1.048.191l11.426 8.426a1.196 1.196 0 0 1 .475 1.191 1.23 1.23 0 0 1-.856.953l-33.279 8.426-.286.049Zm-5.808-9.38 6.189 6.856 30.136-7.665-8.95-6.572-27.375 7.381Z"
-            fill="#210E10"
-         />
-         {/* light shade swatch  */}
-         <path d="m406.399 488.235 37.614-10.713-12.57-9.236-36.564 7.617 11.52 12.332Z" fill="#B682FF" />
-         <path
-            d="m406.019 489.566-13.425-14.377 39.135-8.142 14.903 10.95-40.613 11.569Zm-8.807-12.948 9.57 10.283 34.564-9.854-10.189-7.476-33.945 7.047Z"
-            fill="#B682FF"
-         />
-         <path
-            d="M380.548 444.862c1.605 0 2.905-.938 2.905-2.095 0-1.157-1.3-2.095-2.905-2.095-1.603 0-2.905.938-2.905 2.095 0 1.157 1.302 2.095 2.905 2.095Z"
-            fill="#fff"
-         />
-         {/* right swatch dot */}
-         <path
-            d="M380.548 446.005c-1.666 0-3.189-.761-3.809-2.047a2.63 2.63 0 0 1 0-2.381 3.68 3.68 0 0 1 2.619-1.952c2.048-.475 4.284.381 5.046 1.952a2.944 2.944 0 0 1 0 2.381 3.926 3.926 0 0 1-1.104 1.233 3.964 3.964 0 0 1-1.513.671c-.402.118-.82.166-1.239.143Zm-1.666-3.094c.238.524 1.334.904 2.333.667.204-.034.4-.11.574-.223.173-.116.318-.268.425-.444a.193.193 0 0 0 .049-.065.198.198 0 0 0 0-.156.175.175 0 0 0-.049-.065c-.238-.524-1.285-.905-2.333-.667a1.406 1.406 0 0 0-.572.223 1.447 1.447 0 0 0-.427.444.286.286 0 0 0 0 .286Z"
-            fill="#210E10"
-         />
-
-         {/* ALBUM */}
-         {/* album fill */}
-         <path d="m150.209 504.612-97.22-54.418 92.649-18.616 97.316 35.898-92.745 37.136Z" fill="#fff" />
-         {/* album border */}
-         <path
-            d="M150.209 505.805a1.343 1.343 0 0 1-.619-.143l-97.173-54.418a1.292 1.292 0 0 1-.619-1.239 1.282 1.282 0 0 1 .953-.999l92.65-18.616c.223-.034.45-.017.666.048l97.316 35.898a1.237 1.237 0 0 1 .762 1.094 1.19 1.19 0 0 1-.762 1.144l-92.745 37.135-.429.096Zm-93.84-55.085 93.936 52.561 89.317-35.754-94.078-34.71-89.175 17.903Z"
-            fill="#210E10"
-         />
-         {/* album filled trapez */}
-         <path d="m153.828 494.235-70.892-39.232 53.847-13.712 42.041 10.664-24.996 42.28Z" fill="#5624C2" />
-         <path
-            d="m154.257 495.854-74.51-41.278 57.037-14.521 43.849 11.14-26.376 44.659Zm-68.13-40.42 67.273 37.183 23.567-39.899-40.183-10.187-50.658 12.903Z"
-            fill="#5624C2"
-         />
-         {/* album transparent trapez mark */}
-         <path
-            d="M147.971 479.475a1.135 1.135 0 0 1-.577-.152 1.115 1.115 0 0 1-.423-.42 1.232 1.232 0 0 1 0-1.286l9.475-15.043a1.274 1.274 0 0 1 .904-.524l37.565-3.858a.855.855 0 0 1 .349-.002.84.84 0 0 1 .317.145l23.044 12.522a1.238 1.238 0 0 1 .572 1.285 1.19 1.19 0 0 1-1.048.905l-70.083 6.428h-.095Zm10.189-15.141-7.904 12.522 63.75-5.808-19.234-10.475-36.612 3.761Z"
-            fill="#210E10"
-         />
-
-         {/* SMALL BUBBLES */}
-         {/* big bubble in the right */}
-         <path
-            d="M459.455 68.343c.26 17.669-13.134 32.291-29.917 32.659-16.783.368-30.598-13.657-30.858-31.327-.261-17.669 13.134-32.291 29.916-32.659 16.783-.368 30.599 13.657 30.859 31.327Z"
-            fill="url(#d)"
-         />
-         {/* small bubble in the right */}
-         <path
-            d="M393.472 65.015c.088 5.973-4.44 10.917-10.114 11.041-5.674.125-10.345-4.617-10.433-10.59-.088-5.974 4.441-10.917 10.114-11.042 5.674-.124 10.345 4.617 10.433 10.59Z"
-            fill="url(#e)"
-            fillOpacity={0.5}
-         />
-         {/* small bubble in the left */}
-         <path
-            d="M81.132 158.501c1.39 7.457-3.218 14.613-10.292 15.984-7.073 1.371-13.934-3.562-15.324-11.018-1.39-7.457 3.218-14.613 10.292-15.984 7.073-1.371 13.934 3.562 15.324 11.018Z"
-            fill="url(#f)"
-            fillOpacity={0.5}
-         />
-         {/* bubble on shirt */}
-         <path
-            d="M145.809 379.997c11.374 5.258 20.892 11.85 15.634 23.223-5.258 11.374-18.741 16.331-30.114 11.073-11.374-5.258-16.331-18.741-11.073-30.114 5.258-11.373 14.18-9.44 25.553-4.182Z"
-            fill="url(#g)"
-         />
-      </g>
-      <defs>
-         <linearGradient id="b" x1={107.021} y1={417.826} x2={167.087} y2={395.572} gradientUnits="userSpaceOnUse">
-            <stop offset={0.161} stopColor="#5624C2" />
-            <stop offset={1} stopColor="#B682FF" />
-         </linearGradient>
-         <linearGradient id="c" x1={107.021} y1={417.826} x2={167.087} y2={395.572} gradientUnits="userSpaceOnUse">
-            <stop offset={0.161} stopColor="#5624C2" />
-            <stop offset={1} stopColor="#B682FF" />
-         </linearGradient>
-         <linearGradient id="d" x1={31.335} y1={37.292} x2={519.464} y2={470.165} gradientUnits="userSpaceOnUse">
-            <stop stopColor="#7136F1" />
-            <stop offset={0.708} stopColor="#B682FF" />
-         </linearGradient>
-         <linearGradient id="e" x1={31.335} y1={37.292} x2={519.464} y2={470.165} gradientUnits="userSpaceOnUse">
-            <stop stopColor="#7136F1" />
-            <stop offset={0.708} stopColor="#B682FF" />
-         </linearGradient>
-         <linearGradient id="f" x1={31.335} y1={37.292} x2={519.464} y2={470.165} gradientUnits="userSpaceOnUse">
-            <stop stopColor="#7136F1" />
-            <stop offset={0.708} stopColor="#B682FF" />
-         </linearGradient>
-         <linearGradient id="g" x1={107.021} y1={417.826} x2={167.087} y2={395.572} gradientUnits="userSpaceOnUse">
-            <stop offset={0.161} stopColor="#5624C2" />
-            <stop offset={1} stopColor="#B682FF" />
-         </linearGradient>
-         <filter
-            id="a"
-            x={0.311}
-            y={0.872}
-            width={536.068}
-            height={524.595}
-            filterUnits="userSpaceOnUse"
-            colorInterpolationFilters="sRGB"
-         >
-            <feFlood floodOpacity={0} result="BackgroundImageFix" />
-            <feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha" />
-            <feOffset dy={3.025} />
-            <feGaussianBlur stdDeviation={8.319} />
-            <feComposite in2="hardAlpha" operator="out" />
-            <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.25 0" />
-            <feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3_1172" />
-            <feBlend in="SourceGraphic" in2="effect1_dropShadow_3_1172" result="shape" />
-         </filter>
-      </defs>
-   </svg>
-)
-
-export default SvgComponent
