--- src/SvgPipe.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-import SvgComponent from './Svg'
-type Props = {}
-
-const SvgPipe = (props: Props) => {
-  return (
-    <div>SvgPipe</div>
-  )
-}
-
-export default SvgPipe
