+++ src/components/SwatchList.tsx
@@ -3,7 +3,6 @@ import { usePaletteContext } from '../hooks/usePaletteContext'
 import { CiUnlock } from 'react-icons/ci'
 import { CiLock } from 'react-icons/ci'
 import { PaletteColorType } from '../services/palette'
-import { useScroll } from '../hooks/useScroll'
 import Swatch from './Swatch'
 type Props = {}
 
@@ -11,39 +10,56 @@ const SwatchList = ({}: Props) => {
    const { palette, setLock } = usePaletteContext()
    const { primary, secondary, tertiary, info, success, warning, danger, neutral } = palette
    const colors = [primary, secondary, tertiary, neutral, neutral]
-   const [isTop, setIsTop] = useState(false)
+   const [itemsToShow, setItemsToShow] = useState<number>(5)
+   const [isTop, setIsTop] = useState<boolean>(false)
    const handleLock = (color: PaletteColorType) => {
       setLock(color)
    }
-   const { elRef, getTop } = useScroll()
-   // reference the swatch list to get position
+
+   // wiewport height
    const swatchListRef = useRef<HTMLDivElement>(null)
-
    const swatchList = swatchListRef.current
-   useEffect(() => {
-      const position = swatchList?.getBoundingClientRect()
+   const position = swatchList?.getBoundingClientRect()
 
-      // if scroll position is at swatch list position, log position
+   // viewport width
+   const [width, setWidth] = useState<number>(window.innerWidth)
+   const handleResize = () => {
+      setWidth(window.innerWidth)
+   }
+
+   useEffect(() => {
+      window.addEventListener('resize', handleResize)
+      return () => {
+         window.removeEventListener('resize', handleResize)
+      }
+   })
+
+   useEffect(() => {
+      console.log(width)
+      if (width < 768) {
+         setItemsToShow(3)
+      } else {
+         setItemsToShow(5)
+      }
+   }, [width])
+
+   useEffect(() => {
       if (position && position?.top < 0) {
          setIsTop(true)
       } else {
          setIsTop(false)
       }
-   })
+   }, [position])
 
    return (
-      <div ref={swatchListRef} className="swatch-list-container " style={{ position: isTop ? 'sticky' : 'sticky' }}>
+      <div ref={swatchListRef} className="swatch-list-container ">
          <ul className="swatch-list rounded-2xl">
-            {colors.map((color, index) => {
+            {colors.slice(0, itemsToShow).map((color, index) => {
                return <Swatch key={index} color={color} />
             })}
          </ul>
          <div className="desc">
             <button>Generate</button>
-            <p style={{ opacity: isTop ? '0' : '1', transition:'.2s' }}>
-               Lorem ipsum dolor sit amet consectetur adipisicing elit. 
-               Voluptatem beatae distinctio ut tempore? Sint
-            </p>
          </div>
       </div>
    )
