+++ src/features/palette/index.ts
@@ -1,0 +1,2 @@
+export * from './palette'
+export * from './PaletteType'
