+++ src/layouts/Hero.tsx
@@ -1,181 +1,112 @@
-import React from 'react'
+import React, { useEffect, useState } from 'react'
 import { usePaletteContext } from '../hooks/usePaletteContext'
-import Waves from '../Waves'
-type Props = {}
+import { PaletteColorType } from '../services/palette/PaletteType'
+import { PaletteColorRole } from '../services/palette/PaletteType'
+import SwatchList from '../components/SwatchList'
+type Props = {
+   scrollPosition: number
+}
 
-const Hero = (props: Props) => {
-   const { palette } = usePaletteContext()
+function Showcase({ scrollPosition }: Props) {
+   const { palette, paletteColors, generatePaletteByStyle, generatePaletteColor } = usePaletteContext()
+   const [colorStyle, setColorStyle] = React.useState<'neon' | 'pastel' | 'earth' | 'jewel'>('pastel')
+   const setSassVariable = (variable: string, value: string) => {
+      const root = document.documentElement
+      root.style.setProperty(variable, value)
+   }
+
+   // generate random color
+   const [isSwatchListSticky, setIsSwatchListSticky] = React.useState<boolean>(false)
+
+   const handleGenerate = () => {
+      const randomColor = '#' + Math.floor(Math.random() * 16777215).toString(16)
+      // setPrimary(new Color({ hex: randomColor }))
+
+      const unlockedColors = paletteColors.filter((color: PaletteColorType) => !color.isLocked)
+      unlockedColors.forEach(color => {
+         generatePaletteColor(colorStyle, color.role as PaletteColorRole)
+         console.log(color.role)
+      })
+      generatePaletteByStyle(colorStyle)
+      // setPrimary(new Color({ hex: randomColor }), )
+   }
+
+   useEffect(() => {
+      setSassVariable('--primary500', palette.primary.shade[500].hex)
+      setSassVariable('--primary100', palette.primary.shade[100].hex)
+      setSassVariable('--primary200', palette.primary.shade[200].hex)
+      setSassVariable('--primary300', palette.primary.shade[300].hex)
+      setSassVariable('--primary400', palette.primary.shade[400].hex)
+      setSassVariable('--primary600', palette.primary.shade[600].hex)
+      setSassVariable('--primary700', palette.primary.shade[700].hex)
+      setSassVariable('--primary800', palette.primary.shade[800].hex)
+      setSassVariable('--primary900', palette.primary.shade[900].hex)
+      setSassVariable('--secondary100', palette.secondary.shade[100].hex)
+      setSassVariable('--secondary200', palette.secondary.shade[200].hex)
+      setSassVariable('--secondary300', palette.secondary.shade[300].hex)
+      setSassVariable('--secondary400', palette.secondary.shade[400].hex)
+      setSassVariable('--secondary500', palette.secondary.shade[500].hex)
+      setSassVariable('--secondary600', palette.secondary.shade[600].hex)
+      setSassVariable('--secondary700', palette.secondary.shade[700].hex)
+      setSassVariable('--secondary800', palette.secondary.shade[800].hex)
+      setSassVariable('--secondary900', palette.secondary.shade[900].hex)
+
+      setSassVariable('--neutral100', palette.neutral.shade[100].hex)
+      setSassVariable('--neutral200', palette.neutral.shade[200].hex)
+      setSassVariable('--neutral300', palette.neutral.shade[300].hex)
+      setSassVariable('--neutral400', palette.neutral.shade[400].hex)
+      setSassVariable('--neutral500', palette.neutral.shade[500].hex)
+      setSassVariable('--neutral600', palette.neutral.shade[600].hex)
+      setSassVariable('--neutral700', palette.neutral.shade[700].hex)
+      setSassVariable('--neutral800', palette.neutral.shade[800].hex)
+      setSassVariable('--neutral900', palette.neutral.shade[900].hex)
+   }, [palette])
+
+   const getShadeHex = (color: PaletteColorType, shade: number) => {
+      return color.shade[shade].hex
+   }
+
+   const textColor = ({ r, g, b }: { r: number; g: number; b: number }) => {
+      var yiq = (r * 299 + g * 587 + b * 114) / 1000
+      return yiq >= 128 ? 'black' : 'white'
+   }
+
    return (
-      <div className="mb-16">
-         <div className="bg-gray-100">
-            <div className="px-4 py-16 mx-auto sm:max-w-xl md:max-w-full lg:max-w-screen-xl md:px-24 lg:px-8 lg:py-20">
-               <div className="max-w-xl mb-10 md:mx-auto sm:text-center lg:max-w-2xl md:mb-12">
-                  <h2 className="max-w-lg mb-6 font-sans text-3xl font-bold leading-none tracking-tight text-gray-900 sm:text-4xl md:mx-auto">
-                     <span className="relative inline-block">
-                        <svg
-                           viewBox="0 0 52 24"
-                           fill="currentColor"
-                           className="absolute top-0 left-0 z-0 hidden w-32 -mt-8 -ml-20 text-gray-400 lg:w-32 lg:-ml-28 lg:-mt-10 sm:block"
-                        >
-                           <defs>
-                              <pattern
-                                 id="dc223fcc-6d72-4ebc-b4ef-abe121034d6e"
-                                 x="0"
-                                 y="0"
-                                 width=".135"
-                                 height=".30"
-                              >
-                                 <circle cx="1" cy="1" r=".7" />
-                              </pattern>
-                           </defs>
-                           <rect
-                              fill="url(#dc223fcc-6d72-4ebc-b4ef-abe121034d6e)"
-                              width="52"
-                              height="24"
-                           />
-                        </svg>
-                        <span className="relative">The</span>
-                     </span>{' '}
-                     quick, brown fox jumps over a lazy dog
-                  </h2>
-                  <p className="text-base text-gray-700 md:text-lg">
-                     Sed ut perspiciatis unde omnis iste natus error sit
-                     voluptatem accusantium doloremque rem aperiam, eaque ipsa
-                     quae.
-                  </p>
-               </div>
-               {/* BUTTON */}
-               <div className="flex items-center sm:justify-center w-full">
-                  <a
-                     style={{
-                        background: palette.primary.shade[500].hex,
-                       
-                     }}
-                     href="#_"
-                     className="relative inline-flex items-center justify-center px-10 py-4 overflow-hidden font-mono font-medium tracking-tighter text-white bg-gray-800 rounded-lg group "
-                  >
-                     <span
-                        style={{
-                           background: palette.primary.shade[600].hex,
-                        }}
-                        className="absolute w-0 h-0 transition-all duration-500 ease-out bg-green-500 rounded-full group-hover:w-56 group-hover:h-56"
-                     ></span>
-                     <span className="relative">Generate</span>
-                  </a>
-               </div>
-               {/* WAVES */}
-               {/* <Waves></Waves> */}
-            </div>
+      <section
+         className="showcase-container "
+         // style={{ position: 'relative', background: getShadeHex(palette.neutral, 500) }}
+         style={{ position: 'relative' }}
+      >
+         <div className="scroll " style={{ position: 'sticky', top: '2rem', left: '2rem', color: 'black' }}>
+            {scrollPosition}
          </div>
-
-         <div className="relative px-4 sm:px-0">
-            <div className="absolute inset-0 bg-gray-100 h-1/2" />
-            <div className="relative grid mx-auto overflow-hidden bg-white divide-y rounded shadow sm:divide-y-0 sm:divide-x sm:max-w-screen-sm sm:grid-cols-3 lg:max-w-screen-md">
-               <div className="inline-block p-8 text-center">
-                  <div
-                     style={{
-                        background: palette.secondary.shade[300].hex,
-                        borderRadius: '50%',
-                     }}
-                     className="flex items-center justify-center w-12 h-12 mx-auto mb-4 rounded-full bg-indigo-50"
-                  >
-                     <svg
-                        style={{
-                           background: palette.secondary.shade[100].hex,
-                           borderRadius: '50%',
-                        }}
-                        className="w-10 h-10 text-deep-purple-accent-400"
-                        stroke={palette.secondary.shade[500].hex}
-                        viewBox="0 0 52 52"
-                     >
-                        <polygon
-                           strokeWidth="3"
-                           strokeLinecap="round"
-                           strokeLinejoin="round"
-                           fill="none"
-                           points="29 13 14 29 25 29 23 39 38 23 27 23"
-                        />
-                     </svg>
-                  </div>
-                  <p
-                     style={{ color: palette.secondary.shade[900].hex }}
-                     className="font-bold tracking-wide text-gray-800"
-                  >
-                     Generate pallete
-                  </p>
+         <div className="hero">
+            <header>
+               <div className="h-container  ">
+                  <h1 style={{ color: palette.primary.shade[500].hex }}>
+                     Generate a Rainb<span style={{ color: palette.secondary.shade[600].hex }}>o</span>w <br /> of
+                     Possibilities
+                  </h1>
                </div>
 
-               <div className="inline-block p-8 text-center">
-                  <div
-                     style={{
-                        background: palette.secondary.shade[300].hex,
-                        borderRadius: '50%',
-                     }}
-                     className="flex items-center justify-center w-12 h-12 mx-auto mb-4 rounded-full bg-indigo-50"
-                  >
-                     <svg
-                        style={{
-                           background: palette.secondary.shade[100].hex,
-                           borderRadius: '50%',
-                        }}
-                        className="w-10 h-10 text-deep-purple-accent-400"
-                        stroke={palette.secondary.shade[500].hex}
-                        viewBox="0 0 52 52"
-                     >
-                        <polygon
-                           strokeWidth="3"
-                           strokeLinecap="round"
-                           strokeLinejoin="round"
-                           fill="none"
-                           points="29 13 14 29 25 29 23 39 38 23 27 23"
-                        />
-                     </svg>
+               <div className="flex">
+                  <div className="btn btn-primary " onClick={handleGenerate}>
+                     Generate
                   </div>
-                  <p
-                     style={{ color: palette.secondary.shade[900].hex }}
-                     className="font-bold tracking-wide text-gray-800"
-                  >
-                     Modify to perfection
-                  </p>
+                  <select onChange={ev => setColorStyle(ev.target.value as 'neon' | 'pastel' | 'earth' | 'jewel')}>
+                     <option value="pastel">Pastel</option>
+                     <option value="jewel">Jewel</option>
+                     <option value="earth">Earthy</option>
+                     <option value="neon">Neon</option>
+                  </select>
                </div>
-
-               <div className="inline-block p-8 text-center">
-                  <div
-                     style={{
-                        background: palette.secondary.shade[300].hex,
-                        borderRadius: '50%',
-                     }}
-                     className="flex items-center justify-center w-12 h-12 mx-auto mb-4 rounded-full bg-indigo-50"
-                  >
-                     <svg
-                        style={{
-                           background: palette.secondary.shade[100].hex,
-                           borderRadius: '50%',
-                        }}
-                        className="w-10 h-10 text-deep-purple-accent-400"
-                        stroke={palette.secondary.shade[500].hex}
-                        viewBox="0 0 52 52"
-                     >
-                        <polygon
-                           strokeWidth="3"
-                           strokeLinecap="round"
-                           strokeLinejoin="round"
-                           fill="none"
-                           points="29 13 14 29 25 29 23 39 38 23 27 23"
-                        />
-                     </svg>
-                  </div>
-                  <p
-                     style={{ color: palette.secondary.shade[900].hex }}
-                     className="font-bold tracking-wide text-gray-800"
-                  >
-                     Use it in your project
-                  </p>
-               </div>
-            </div>
+            </header>
+            <div className="img"></div>
          </div>
-      </div>
+         <SwatchList></SwatchList>
+         <div style={{ position: 'absolute', bottom: 0, right: 0, width: '100%' }}>{/* <Waves /> */}</div>
+      </section>
    )
 }
 
-export default Hero
+export default Showcase
